World Bank country indicator data

List all country indicators supported by the World Bank API.

Usage

wb_country_indicator(
  indicator = "NY.GDP.MKTP.CD",
  country = NULL,
  lang = "en",
  start_date = NULL,
  end_date = NULL
)

Source

https://api.worldbank.org/v2/country/{country}/indicator/{indicator}

Arguments

indicator

(character()) indicators to query.

country

(character()) countries to query. Default NULL. If NULL, all countries are returned.

lang

(character(1)) language to query. Default "en".

start_date

(character(1) | integer(1)) start date to query. Default NULL. Supported formats:

YYYY for yearly data (e.g. 2020 or "2020")

YYYYQ[1-4] for quarterly data (e.g. "2020Q1")

YYYYM[1-12] for monthly data (e.g. "2020M02")

end_date

(character(1) | integer(1)) end date to query, in the same format as start_date. Default NULL.

Value

A data.frame() with the available country indicators. The columns are:

date

The date

indicator_id

The indicator ID.

indicator_name

The indicator name.

country_id

The country ID.

country_name

The country name.

country_code

The country code.

value

The indicator value.

unit

The indicator unit.

obs_status

The observation status.

decimal

The decimal.

Examples

wb_country_indicator("NY.GDP.MKTP.CD", "US")
#>    date   indicator_id    indicator_name country_id  country_name country_code
#> 1  2023 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 2  2022 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 3  2021 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 4  2020 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 5  2019 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 6  2018 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 7  2017 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 8  2016 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 9  2015 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 10 2014 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 11 2013 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 12 2012 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 13 2011 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 14 2010 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 15 2009 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 16 2008 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 17 2007 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 18 2006 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 19 2005 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 20 2004 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 21 2003 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 22 2002 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 23 2001 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 24 2000 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 25 1999 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 26 1998 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 27 1997 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 28 1996 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 29 1995 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 30 1994 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 31 1993 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 32 1992 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 33 1991 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 34 1990 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 35 1989 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 36 1988 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 37 1987 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 38 1986 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 39 1985 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 40 1984 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 41 1983 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 42 1982 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 43 1981 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 44 1980 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 45 1979 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 46 1978 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 47 1977 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 48 1976 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 49 1975 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 50 1974 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 51 1973 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 52 1972 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 53 1971 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 54 1970 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 55 1969 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 56 1968 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 57 1967 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 58 1966 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 59 1965 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 60 1964 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 61 1963 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 62 1962 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 63 1961 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 64 1960 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#>           value unit obs_status decimal
#> 1  2.772071e+13 <NA>       <NA>       0
#> 2  2.600689e+13 <NA>       <NA>       0
#> 3  2.368117e+13 <NA>       <NA>       0
#> 4  2.135410e+13 <NA>       <NA>       0
#> 5  2.153998e+13 <NA>       <NA>       0
#> 6  2.065652e+13 <NA>       <NA>       0
#> 7  1.961210e+13 <NA>       <NA>       0
#> 8  1.880491e+13 <NA>       <NA>       0
#> 9  1.829502e+13 <NA>       <NA>       0
#> 10 1.760814e+13 <NA>       <NA>       0
#> 11 1.688068e+13 <NA>       <NA>       0
#> 12 1.625397e+13 <NA>       <NA>       0
#> 13 1.559973e+13 <NA>       <NA>       0
#> 14 1.504897e+13 <NA>       <NA>       0
#> 15 1.447807e+13 <NA>       <NA>       0
#> 16 1.476986e+13 <NA>       <NA>       0
#> 17 1.447423e+13 <NA>       <NA>       0
#> 18 1.381558e+13 <NA>       <NA>       0
#> 19 1.303920e+13 <NA>       <NA>       0
#> 20 1.221720e+13 <NA>       <NA>       0
#> 21 1.145645e+13 <NA>       <NA>       0
#> 22 1.092911e+13 <NA>       <NA>       0
#> 23 1.058193e+13 <NA>       <NA>       0
#> 24 1.025095e+13 <NA>       <NA>       0
#> 25 9.631172e+12 <NA>       <NA>       0
#> 26 9.062817e+12 <NA>       <NA>       0
#> 27 8.577552e+12 <NA>       <NA>       0
#> 28 8.073122e+12 <NA>       <NA>       0
#> 29 7.639749e+12 <NA>       <NA>       0
#> 30 7.287236e+12 <NA>       <NA>       0
#> 31 6.858559e+12 <NA>       <NA>       0
#> 32 6.520327e+12 <NA>       <NA>       0
#> 33 6.158129e+12 <NA>       <NA>       0
#> 34 5.963144e+12 <NA>       <NA>       0
#> 35 5.641580e+12 <NA>       <NA>       0
#> 36 5.236438e+12 <NA>       <NA>       0
#> 37 4.855215e+12 <NA>       <NA>       0
#> 38 4.579631e+12 <NA>       <NA>       0
#> 39 4.338979e+12 <NA>       <NA>       0
#> 40 4.037613e+12 <NA>       <NA>       0
#> 41 3.634038e+12 <NA>       <NA>       0
#> 42 3.343789e+12 <NA>       <NA>       0
#> 43 3.207041e+12 <NA>       <NA>       0
#> 44 2.857307e+12 <NA>       <NA>       0
#> 45 2.627333e+12 <NA>       <NA>       0
#> 46 2.351599e+12 <NA>       <NA>       0
#> 47 2.081826e+12 <NA>       <NA>       0
#> 48 1.873412e+12 <NA>       <NA>       0
#> 49 1.684904e+12 <NA>       <NA>       0
#> 50 1.545243e+12 <NA>       <NA>       0
#> 51 1.425376e+12 <NA>       <NA>       0
#> 52 1.279110e+12 <NA>       <NA>       0
#> 53 1.164850e+12 <NA>       <NA>       0
#> 54 1.073303e+12 <NA>       <NA>       0
#> 55 1.017438e+12 <NA>       <NA>       0
#> 56 9.402250e+11 <NA>       <NA>       0
#> 57 8.596200e+11 <NA>       <NA>       0
#> 58 8.130328e+11 <NA>       <NA>       0
#> 59 7.419049e+11 <NA>       <NA>       0
#> 60 6.841446e+11 <NA>       <NA>       0
#> 61 6.370586e+11 <NA>       <NA>       0
#> 62 6.036394e+11 <NA>       <NA>       0
#> 63 5.619403e+11 <NA>       <NA>       0
#> 64 5.419886e+11 <NA>       <NA>       0
wb_country_indicator(
  "DPANUSSPB", c("CHN", "BRA"),
  start_date = "2012M01", end_date = "2012M08"
)
#>    date indicator_id
#> 1  2025    DPANUSSPB
#> 2  2024    DPANUSSPB
#> 3  2023    DPANUSSPB
#> 4  2022    DPANUSSPB
#> 5  2021    DPANUSSPB
#> 6  2020    DPANUSSPB
#> 7  2019    DPANUSSPB
#> 8  2018    DPANUSSPB
#> 9  2017    DPANUSSPB
#> 10 2016    DPANUSSPB
#> 11 2015    DPANUSSPB
#> 12 2014    DPANUSSPB
#> 13 2013    DPANUSSPB
#> 14 2012    DPANUSSPB
#> 15 2011    DPANUSSPB
#> 16 2010    DPANUSSPB
#> 17 2009    DPANUSSPB
#> 18 2008    DPANUSSPB
#> 19 2007    DPANUSSPB
#> 20 2006    DPANUSSPB
#> 21 2005    DPANUSSPB
#> 22 2004    DPANUSSPB
#> 23 2003    DPANUSSPB
#> 24 2002    DPANUSSPB
#> 25 2001    DPANUSSPB
#> 26 2000    DPANUSSPB
#> 27 1999    DPANUSSPB
#> 28 1998    DPANUSSPB
#> 29 1997    DPANUSSPB
#> 30 1996    DPANUSSPB
#> 31 1995    DPANUSSPB
#> 32 1994    DPANUSSPB
#> 33 1993    DPANUSSPB
#> 34 1992    DPANUSSPB
#> 35 1991    DPANUSSPB
#> 36 1990    DPANUSSPB
#> 37 2025    DPANUSSPB
#> 38 2024    DPANUSSPB
#> 39 2023    DPANUSSPB
#> 40 2022    DPANUSSPB
#> 41 2021    DPANUSSPB
#> 42 2020    DPANUSSPB
#> 43 2019    DPANUSSPB
#> 44 2018    DPANUSSPB
#> 45 2017    DPANUSSPB
#> 46 2016    DPANUSSPB
#> 47 2015    DPANUSSPB
#> 48 2014    DPANUSSPB
#> 49 2013    DPANUSSPB
#> 50 2012    DPANUSSPB
#> 51 2011    DPANUSSPB
#> 52 2010    DPANUSSPB
#> 53 2009    DPANUSSPB
#> 54 2008    DPANUSSPB
#> 55 2007    DPANUSSPB
#> 56 2006    DPANUSSPB
#> 57 2005    DPANUSSPB
#> 58 2004    DPANUSSPB
#> 59 2003    DPANUSSPB
#> 60 2002    DPANUSSPB
#> 61 2001    DPANUSSPB
#> 62 2000    DPANUSSPB
#> 63 1999    DPANUSSPB
#> 64 1998    DPANUSSPB
#> 65 1997    DPANUSSPB
#> 66 1996    DPANUSSPB
#> 67 1995    DPANUSSPB
#> 68 1994    DPANUSSPB
#> 69 1993    DPANUSSPB
#> 70 1992    DPANUSSPB
#> 71 1991    DPANUSSPB
#> 72 1990    DPANUSSPB
#> 73 1989    DPANUSSPB
#> 74 1988    DPANUSSPB
#> 75 1987    DPANUSSPB
#>                                                        indicator_name
#> 1  Exchange rate, new LCU per USD extended backward, period average,,
#> 2  Exchange rate, new LCU per USD extended backward, period average,,
#> 3  Exchange rate, new LCU per USD extended backward, period average,,
#> 4  Exchange rate, new LCU per USD extended backward, period average,,
#> 5  Exchange rate, new LCU per USD extended backward, period average,,
#> 6  Exchange rate, new LCU per USD extended backward, period average,,
#> 7  Exchange rate, new LCU per USD extended backward, period average,,
#> 8  Exchange rate, new LCU per USD extended backward, period average,,
#> 9  Exchange rate, new LCU per USD extended backward, period average,,
#> 10 Exchange rate, new LCU per USD extended backward, period average,,
#> 11 Exchange rate, new LCU per USD extended backward, period average,,
#> 12 Exchange rate, new LCU per USD extended backward, period average,,
#> 13 Exchange rate, new LCU per USD extended backward, period average,,
#> 14 Exchange rate, new LCU per USD extended backward, period average,,
#> 15 Exchange rate, new LCU per USD extended backward, period average,,
#> 16 Exchange rate, new LCU per USD extended backward, period average,,
#> 17 Exchange rate, new LCU per USD extended backward, period average,,
#> 18 Exchange rate, new LCU per USD extended backward, period average,,
#> 19 Exchange rate, new LCU per USD extended backward, period average,,
#> 20 Exchange rate, new LCU per USD extended backward, period average,,
#> 21 Exchange rate, new LCU per USD extended backward, period average,,
#> 22 Exchange rate, new LCU per USD extended backward, period average,,
#> 23 Exchange rate, new LCU per USD extended backward, period average,,
#> 24 Exchange rate, new LCU per USD extended backward, period average,,
#> 25 Exchange rate, new LCU per USD extended backward, period average,,
#> 26 Exchange rate, new LCU per USD extended backward, period average,,
#> 27 Exchange rate, new LCU per USD extended backward, period average,,
#> 28 Exchange rate, new LCU per USD extended backward, period average,,
#> 29 Exchange rate, new LCU per USD extended backward, period average,,
#> 30 Exchange rate, new LCU per USD extended backward, period average,,
#> 31 Exchange rate, new LCU per USD extended backward, period average,,
#> 32 Exchange rate, new LCU per USD extended backward, period average,,
#> 33 Exchange rate, new LCU per USD extended backward, period average,,
#> 34 Exchange rate, new LCU per USD extended backward, period average,,
#> 35 Exchange rate, new LCU per USD extended backward, period average,,
#> 36 Exchange rate, new LCU per USD extended backward, period average,,
#> 37 Exchange rate, new LCU per USD extended backward, period average,,
#> 38 Exchange rate, new LCU per USD extended backward, period average,,
#> 39 Exchange rate, new LCU per USD extended backward, period average,,
#> 40 Exchange rate, new LCU per USD extended backward, period average,,
#> 41 Exchange rate, new LCU per USD extended backward, period average,,
#> 42 Exchange rate, new LCU per USD extended backward, period average,,
#> 43 Exchange rate, new LCU per USD extended backward, period average,,
#> 44 Exchange rate, new LCU per USD extended backward, period average,,
#> 45 Exchange rate, new LCU per USD extended backward, period average,,
#> 46 Exchange rate, new LCU per USD extended backward, period average,,
#> 47 Exchange rate, new LCU per USD extended backward, period average,,
#> 48 Exchange rate, new LCU per USD extended backward, period average,,
#> 49 Exchange rate, new LCU per USD extended backward, period average,,
#> 50 Exchange rate, new LCU per USD extended backward, period average,,
#> 51 Exchange rate, new LCU per USD extended backward, period average,,
#> 52 Exchange rate, new LCU per USD extended backward, period average,,
#> 53 Exchange rate, new LCU per USD extended backward, period average,,
#> 54 Exchange rate, new LCU per USD extended backward, period average,,
#> 55 Exchange rate, new LCU per USD extended backward, period average,,
#> 56 Exchange rate, new LCU per USD extended backward, period average,,
#> 57 Exchange rate, new LCU per USD extended backward, period average,,
#> 58 Exchange rate, new LCU per USD extended backward, period average,,
#> 59 Exchange rate, new LCU per USD extended backward, period average,,
#> 60 Exchange rate, new LCU per USD extended backward, period average,,
#> 61 Exchange rate, new LCU per USD extended backward, period average,,
#> 62 Exchange rate, new LCU per USD extended backward, period average,,
#> 63 Exchange rate, new LCU per USD extended backward, period average,,
#> 64 Exchange rate, new LCU per USD extended backward, period average,,
#> 65 Exchange rate, new LCU per USD extended backward, period average,,
#> 66 Exchange rate, new LCU per USD extended backward, period average,,
#> 67 Exchange rate, new LCU per USD extended backward, period average,,
#> 68 Exchange rate, new LCU per USD extended backward, period average,,
#> 69 Exchange rate, new LCU per USD extended backward, period average,,
#> 70 Exchange rate, new LCU per USD extended backward, period average,,
#> 71 Exchange rate, new LCU per USD extended backward, period average,,
#> 72 Exchange rate, new LCU per USD extended backward, period average,,
#> 73 Exchange rate, new LCU per USD extended backward, period average,,
#> 74 Exchange rate, new LCU per USD extended backward, period average,,
#> 75 Exchange rate, new LCU per USD extended backward, period average,,
#>    country_id country_name country_code       value unit obs_status decimal
#> 1         BRA       Brazil         <NA> 6.079468750 <NA>       <NA>       0
#> 2         BRA       Brazil         <NA> 5.392557392 <NA>       <NA>       0
#> 3         BRA       Brazil         <NA> 4.994074475 <NA>       <NA>       0
#> 4         BRA       Brazil         <NA> 5.165223848 <NA>       <NA>       0
#> 5         BRA       Brazil         <NA> 5.396087280 <NA>       <NA>       0
#> 6         BRA       Brazil         <NA> 5.157586594 <NA>       <NA>       0
#> 7         BRA       Brazil         <NA> 3.944611075 <NA>       <NA>       0
#> 8         BRA       Brazil         <NA> 3.654127300 <NA>       <NA>       0
#> 9         BRA       Brazil         <NA> 3.190255448 <NA>       <NA>       0
#> 10        BRA       Brazil         <NA> 3.486714603 <NA>       <NA>       0
#> 11        BRA       Brazil         <NA> 3.331694320 <NA>       <NA>       0
#> 12        BRA       Brazil         <NA> 2.354272770 <NA>       <NA>       0
#> 13        BRA       Brazil         <NA> 2.158732053 <NA>       <NA>       0
#> 14        BRA       Brazil         <NA> 1.954158795 <NA>       <NA>       0
#> 15        BRA       Brazil         <NA> 1.674466152 <NA>       <NA>       0
#> 16        BRA       Brazil         <NA> 1.760206488 <NA>       <NA>       0
#> 17        BRA       Brazil         <NA> 1.999489729 <NA>       <NA>       0
#> 18        BRA       Brazil         <NA> 1.836026973 <NA>       <NA>       0
#> 19        BRA       Brazil         <NA> 1.947450509 <NA>       <NA>       0
#> 20        BRA       Brazil         <NA> 2.175136433 <NA>       <NA>       0
#> 21        BRA       Brazil         <NA> 2.435280782 <NA>       <NA>       0
#> 22        BRA       Brazil         <NA> 2.926277840 <NA>       <NA>       0
#> 23        BRA       Brazil         <NA> 3.077288540 <NA>       <NA>       0
#> 24        BRA       Brazil         <NA> 2.921882492 <NA>       <NA>       0
#> 25        BRA       Brazil         <NA> 2.350970797 <NA>       <NA>       0
#> 26        BRA       Brazil         <NA> 1.829968786 <NA>       <NA>       0
#> 27        BRA       Brazil         <NA> 1.814313261 <NA>       <NA>       0
#> 28        BRA       Brazil         <NA> 1.160159280 <NA>       <NA>       0
#> 29        BRA       Brazil         <NA> 1.077677409 <NA>       <NA>       0
#> 30        BRA       Brazil         <NA> 1.004733893 <NA>       <NA>       0
#> 31        BRA       Brazil         <NA> 0.916864239 <NA>       <NA>       0
#> 32        BRA       Brazil         <NA> 0.638294927 <NA>       <NA>       0
#> 33        BRA       Brazil         <NA> 0.032091361 <NA>       <NA>       0
#> 34        BRA       Brazil         <NA> 0.001787298 <NA>       <NA>       0
#> 35        BRA       Brazil         <NA> 0.000350000 <NA>       <NA>       0
#> 36        BRA       Brazil         <NA> 0.000200000 <NA>       <NA>       0
#> 37        CHN        China         <NA> 7.316764706 <NA>       <NA>       0
#> 38        CHN        China         <NA> 7.197247540 <NA>       <NA>       0
#> 39        CHN        China         <NA> 7.080443704 <NA>       <NA>       0
#> 40        CHN        China         <NA> 6.729639518 <NA>       <NA>       0
#> 41        CHN        China         <NA> 6.449797254 <NA>       <NA>       0
#> 42        CHN        China         <NA> 6.902117964 <NA>       <NA>       0
#> 43        CHN        China         <NA> 6.909922683 <NA>       <NA>       0
#> 44        CHN        China         <NA> 6.613852907 <NA>       <NA>       0
#> 45        CHN        China         <NA> 6.757446433 <NA>       <NA>       0
#> 46        CHN        China         <NA> 6.645067075 <NA>       <NA>       0
#> 47        CHN        China         <NA> 6.285358703 <NA>       <NA>       0
#> 48        CHN        China         <NA> 6.161507718 <NA>       <NA>       0
#> 49        CHN        China         <NA> 6.148686634 <NA>       <NA>       0
#> 50        CHN        China         <NA> 6.308628988 <NA>       <NA>       0
#> 51        CHN        China         <NA> 6.464558612 <NA>       <NA>       0
#> 52        CHN        China         <NA> 6.769090446 <NA>       <NA>       0
#> 53        CHN        China         <NA> 6.831505565 <NA>       <NA>       0
#> 54        CHN        China         <NA> 6.950109677 <NA>       <NA>       0
#> 55        CHN        China         <NA> 7.606216150 <NA>       <NA>       0
#> 56        CHN        China         <NA> 7.972375478 <NA>       <NA>       0
#> 57        CHN        China         <NA> 8.193666329 <NA>       <NA>       0
#> 58        CHN        China         <NA> 8.276925333 <NA>       <NA>       0
#> 59        CHN        China         <NA> 8.277098414 <NA>       <NA>       0
#> 60        CHN        China         <NA> 8.277018989 <NA>       <NA>       0
#> 61        CHN        China         <NA> 8.277149155 <NA>       <NA>       0
#> 62        CHN        China         <NA> 8.278492082 <NA>       <NA>       0
#> 63        CHN        China         <NA> 8.278342026 <NA>       <NA>       0
#> 64        CHN        China         <NA> 8.279089770 <NA>       <NA>       0
#> 65        CHN        China         <NA> 8.289856415 <NA>       <NA>       0
#> 66        CHN        China         <NA> 8.314231090 <NA>       <NA>       0
#> 67        CHN        China         <NA> 8.351316282 <NA>       <NA>       0
#> 68        CHN        China         <NA> 8.610021425 <NA>       <NA>       0
#> 69        CHN        China         <NA> 5.789434775 <NA>       <NA>       0
#> 70        CHN        China         <NA> 5.528120069 <NA>       <NA>       0
#> 71        CHN        China         <NA> 5.334966938 <NA>       <NA>       0
#> 72        CHN        China         <NA> 4.794740368 <NA>       <NA>       0
#> 73        CHN        China         <NA> 3.770805375 <NA>       <NA>       0
#> 74        CHN        China         <NA> 3.731000000 <NA>       <NA>       0
#> 75        CHN        China         <NA> 3.731000000 <NA>       <NA>       0