World Bank source data

List all sources supported by the World Bank API.

Usage

wb_source(source = NULL, lang = "en")

Source

https://api.worldbank.org/v2/sources

Arguments

source

character() source to query. Default NULL. If NULL, all sources are returned.

lang

character(1) language to query. Default "en".

Value

A data.frame() with the available sources. The columns are:

id

The source ID.

last_updated

The date the source was last updated.

name

The source name.

code

The source code.

description

The source description.

url

The source URL.

data_availability

Whether the source has data available.

metadata_availability

Whether the source has metadata available.

concepts

The concepts associated with the source.

See also

Examples

wb_source()
#>    id last_updated
#> 1   1   2021-08-18
#> 2   2   2024-11-13
#> 3   3   2024-11-05
#> 4   5   2016-03-21
#> 5   6   2024-12-03
#> 6  11   2013-02-22
#> 7  12   2024-06-25
#> 8  13   2022-03-25
#> 9  14   2024-09-24
#> 10 15   2024-12-04
#> 11 16   2024-07-01
#> 12 18   2021-07-23
#> 13 19   2018-09-19
#> 14 20   2024-11-19
#> 15 22   2024-10-31
#> 16 23   2024-10-31
#> 17 25   2024-11-13
#> 18 27   2024-06-11
#> 19 28   2024-03-28
#> 20 29   2024-02-13
#> 21 30   2016-03-31
#> 22 31   2024-06-28
#> 23 32   2022-09-23
#> 24 33   2019-02-27
#> 25 34   2013-04-12
#> 26 35   2018-06-30
#> 27 37   2016-04-06
#> 28 38   2015-04-27
#> 29 39   2022-12-13
#> 30 40   2024-07-01
#> 31 41   2015-05-22
#> 32 43   2023-06-08
#> 33 45   2022-08-11
#> 34 46   2022-07-22
#> 35 50   2017-09-21
#> 36 54   2024-11-11
#> 37 57   2023-08-10
#> 38 58   2019-04-25
#> 39 59   2024-10-08
#> 40 60   2018-04-19
#> 41 61   2019-04-25
#> 42 62   2020-08-24
#> 43 63   2020-09-21
#> 44 64   2022-09-20
#> 45 65   2019-11-06
#> 46 66   2024-03-26
#> 47 67   2024-12-04
#> 48 68   2024-12-04
#> 49 69   2019-06-27
#> 50 70   2024-11-30
#> 51 71   2021-09-10
#> 52 73   2019-06-27
#> 53 75   2024-04-16
#> 54 76   2020-12-07
#> 55 77   2020-12-07
#> 56 78   2020-10-21
#> 57 79   2024-07-30
#> 58 80   2020-07-25
#> 59 81   2024-12-03
#> 60 82   2021-03-24
#> 61 83   2024-11-11
#> 62 84   2024-10-03
#> 63 85   2024-04-29
#> 64 86   2023-05-22
#> 65 87   2022-10-07
#> 66 88   2024-07-24
#> 67 89   2024-12-05
#> 68 90   2024-08-04
#> 69 91   2024-08-19
#>                                                                    name code
#> 1                                                        Doing Business  DBS
#> 2                                          World Development Indicators  WDI
#> 3                                       Worldwide Governance Indicators  WGI
#> 4                                     Subnational Malnutrition Database  SNM
#> 5                                         International Debt Statistics  IDS
#> 6                                         Africa Development Indicators  ADI
#> 7                                                  Education Statistics  EDS
#> 8                                                    Enterprise Surveys  ESY
#> 9                                                     Gender Statistics  GDS
#> 10                                              Global Economic Monitor  GEM
#> 11                           Health Nutrition and Population Statistics  HNP
#> 12                                       IDA Results Measurement System  IDA
#> 13                                         Millennium Development Goals  MDG
#> 14                                         Quarterly Public Sector Debt  PSD
#> 15                              Quarterly External Debt Statistics SDDS  QDS
#> 16                              Quarterly External Debt Statistics GDDS  QDG
#> 17                                                                 Jobs  JOB
#> 18                                            Global Economic Prospects  GEP
#> 19                                           Global Financial Inclusion  FDX
#> 20  The Atlas of Social Protection: Indicators of Resilience and Equity  GSP
#> 21        Exporter Dynamics Database – Indicators at Country-Year Level  ED1
#> 22                          Country Policy and Institutional Assessment  CPI
#> 23                                         Global Financial Development  GFD
#> 24                                   G20 Financial Inclusion Indicators  G2F
#> 25                                     Global Partnership for Education  GPE
#> 26                                           Sustainable Energy for All  SE4
#> 27                                                       LAC Equity Lab  LEL
#> 28                                                  Subnational Poverty  SNP
#> 29        Health Nutrition and Population Statistics by Wealth Quintile  HNQ
#> 30                                 Population estimates and projections  HPP
#> 31                 Country Partnership Strategy for India (FY2013 - 17)  CPS
#> 32                                                 Adjusted Net Savings  WAT
#> 33                  Indonesia Database for Policy and Economic Research  IDD
#> 34                                        Sustainable Development Goals  SDG
#> 35                                               Subnational Population  SNT
#> 36                                              Joint External Debt Hub  JED
#> 37                                                WDI Database Archives  WDA
#> 38                                            Universal Health Coverage  UHC
#> 39                                                      Wealth Accounts  WAC
#> 40                                                     Economic Fitness  EFT
#> 41                                              PPPs Regulatory Quality  PRQ
#> 42                          International Comparison Program (ICP) 2011  ICP
#> 43                                                  Human Capital Index  HCI
#> 44                                     Worldwide Bureaucracy Indicators  WBI
#> 45                    Health Equity and Financial Protection Indicators  HPI
#> 46                                          Logistics Performance Index  LPI
#> 47                                                            PEFA 2011  PF1
#> 48                                                            PEFA 2016  PF6
#> 49            Global Financial Inclusion and Consumer Protection Survey  RFA
#> 50                                                   Economic Fitness 2  EF2
#> 51                          International Comparison Program (ICP) 2005  IC5
#> 52 Global Financial Inclusion and Consumer Protection Survey (Internal)  RFI
#> 53                        Environment, Social and Governance (ESG) Data  ESG
#> 54                      Remittance Prices Worldwide (Sending Countries)  RWS
#> 55                    Remittance Prices Worldwide (Receiving Countries)  RWR
#> 56                                                             ICP 2017  IC7
#> 57                                                           PEFA_GRPFM  GRP
#> 58                           Gender Disaggregated Labor Database (GDLD)  GDL
#> 59                                  International Debt Statistics: DSSI  DSI
#> 60                                            Global Public Procurement  GPP
#> 61                             Statistical Performance Indicators (SPI)  SPI
#> 62                                                     Education Policy  EDP
#> 63                                                        PEFA_2021_SNG  SNG
#> 64                               Global Jobs Indicators Database (JOIN)  JON
#> 65                        Country Climate and Development Report (CCDR)  CCD
#> 66                                            Food Prices for Nutrition  FPN
#> 67                           Identification for Development (ID4D) Data  ID4
#> 68                                                             ICP 2021  IC2
#> 69                                                           PEFA_CRPFM  CRP
#>    description  url data_availability metadata_availability concepts
#> 1         <NA> <NA>              TRUE                  TRUE        3
#> 2         <NA> <NA>              TRUE                  TRUE        3
#> 3         <NA> <NA>              TRUE                  TRUE        3
#> 4         <NA> <NA>              TRUE                  TRUE        3
#> 5         <NA> <NA>              TRUE                  TRUE        4
#> 6         <NA> <NA>              TRUE                  TRUE        3
#> 7         <NA> <NA>              TRUE                  TRUE        3
#> 8         <NA> <NA>              TRUE                  TRUE        3
#> 9         <NA> <NA>              TRUE                  TRUE        3
#> 10        <NA> <NA>              TRUE                  TRUE        3
#> 11        <NA> <NA>              TRUE                  TRUE        3
#> 12        <NA> <NA>              TRUE                  TRUE        3
#> 13        <NA> <NA>              TRUE                  TRUE        3
#> 14        <NA> <NA>              TRUE                  TRUE        3
#> 15        <NA> <NA>              TRUE                  TRUE        3
#> 16        <NA> <NA>              TRUE                  TRUE        3
#> 17        <NA> <NA>              TRUE                  TRUE        3
#> 18        <NA> <NA>              TRUE                 FALSE        3
#> 19        <NA> <NA>              TRUE                  TRUE        3
#> 20        <NA> <NA>              TRUE                  TRUE        3
#> 21        <NA> <NA>              TRUE                 FALSE        3
#> 22        <NA> <NA>              TRUE                  TRUE        3
#> 23        <NA> <NA>              TRUE                 FALSE        3
#> 24        <NA> <NA>              TRUE                  TRUE        3
#> 25        <NA> <NA>              TRUE                  TRUE        3
#> 26        <NA> <NA>              TRUE                  TRUE        3
#> 27        <NA> <NA>              TRUE                  TRUE        3
#> 28        <NA> <NA>              TRUE                 FALSE        3
#> 29        <NA> <NA>              TRUE                  TRUE        3
#> 30        <NA> <NA>              TRUE                  TRUE        3
#> 31        <NA> <NA>              TRUE                 FALSE        3
#> 32        <NA> <NA>              TRUE                 FALSE        3
#> 33        <NA> <NA>              TRUE                 FALSE        3
#> 34        <NA> <NA>              TRUE                 FALSE        3
#> 35        <NA> <NA>              TRUE                 FALSE        3
#> 36        <NA> <NA>              TRUE                 FALSE        3
#> 37        <NA> <NA>              TRUE                  TRUE        4
#> 38        <NA> <NA>              TRUE                  TRUE        3
#> 39        <NA> <NA>              TRUE                  TRUE        3
#> 40        <NA> <NA>              TRUE                  TRUE        3
#> 41        <NA> <NA>              TRUE                  TRUE        3
#> 42        <NA> <NA>             FALSE                  TRUE        4
#> 43        <NA> <NA>              TRUE                  TRUE        3
#> 44        <NA> <NA>              TRUE                 FALSE        3
#> 45        <NA> <NA>              TRUE                  TRUE        3
#> 46        <NA> <NA>              TRUE                  TRUE        3
#> 47        <NA> <NA>              TRUE                  TRUE        4
#> 48        <NA> <NA>              TRUE                  TRUE        4
#> 49        <NA> <NA>              TRUE                  TRUE        3
#> 50        <NA> <NA>              TRUE                  TRUE        3
#> 51        <NA> <NA>              TRUE                 FALSE        4
#> 52        <NA> <NA>              TRUE                  TRUE        3
#> 53        <NA> <NA>              TRUE                  TRUE        3
#> 54        <NA> <NA>              TRUE                 FALSE        3
#> 55        <NA> <NA>              TRUE                 FALSE        3
#> 56        <NA> <NA>              TRUE                  TRUE        4
#> 57        <NA> <NA>              TRUE                  TRUE        4
#> 58        <NA> <NA>              TRUE                 FALSE        4
#> 59        <NA> <NA>              TRUE                 FALSE        4
#> 60        <NA> <NA>              TRUE                 FALSE        3
#> 61        <NA> <NA>              TRUE                  TRUE        3
#> 62        <NA> <NA>              TRUE                  TRUE        3
#> 63        <NA> <NA>              TRUE                 FALSE        4
#> 64        <NA> <NA>              TRUE                  TRUE        3
#> 65        <NA> <NA>              TRUE                  TRUE        3
#> 66        <NA> <NA>              TRUE                  TRUE        4
#> 67        <NA> <NA>              TRUE                  TRUE        3
#> 68        <NA> <NA>              TRUE                  TRUE        4
#> 69        <NA> <NA>              TRUE                  TRUE        4