Return main poverty and inequality statistics

Return main poverty and inequality statistics

Usage

pip_data(
  country = NULL,
  year = NULL,
  povline = 2.15,
  popshare = NULL,
  fill_gaps = FALSE,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  additional_ind = FALSE,
  release_version = NULL,
  ppp_version = NULL,
  version = NULL
)

Source

https://pip.worldbank.org/api

Arguments

country

(character()) countries for which statistics are to be computed, specified as ISO3 codes. Default NULL.

year

(character() | numeric()) year(s) for which statistics are to be computed, specified as YYYY. Default NULL.

povline

(numeric(1)) poverty line to be used to compute poverty mesures. Poverty lines are only accepted up to 3 decimals. Default 2.15.

popshare

(numeric(1)) proportion of the population living below the poverty line. Will be ignored if povline is specified. Default NULL.

fill_gaps

(logical(1)) whether to fill gaps in the data. Default FALSE.

welfare_type

(character(1)) type of welfare measure to be used. Default "all".

reporting_level

(character(1)) level of reporting for the statistics. Default "all".

additional_ind

(logical(1)) whether to include additional indicators. Default FALSE.

release_version

(character(1)) version of the data release in YYYYMMDD format. Default NULL.

ppp_version

(character(1) | numeric(1)) version of the data. Default NULL.

version

(character(1)) version of the data. Default NULL.

Value

A data.frame() with the requested statistics.

See also

Examples

# \donttest{
pip_data(c("ZAF", "ZMB"))
#>           region_name region_code country_name country_code reporting_year
#> 1  Sub-Saharan Africa         SSA South Africa          ZAF           1993
#> 2  Sub-Saharan Africa         SSA South Africa          ZAF           2000
#> 3  Sub-Saharan Africa         SSA South Africa          ZAF           2005
#> 4  Sub-Saharan Africa         SSA South Africa          ZAF           2008
#> 5  Sub-Saharan Africa         SSA South Africa          ZAF           2010
#> 6  Sub-Saharan Africa         SSA South Africa          ZAF           2014
#> 7  Sub-Saharan Africa         SSA       Zambia          ZMB           1991
#> 8  Sub-Saharan Africa         SSA       Zambia          ZMB           1993
#> 9  Sub-Saharan Africa         SSA       Zambia          ZMB           1996
#> 10 Sub-Saharan Africa         SSA       Zambia          ZMB           1998
#> 11 Sub-Saharan Africa         SSA       Zambia          ZMB           2002
#> 12 Sub-Saharan Africa         SSA       Zambia          ZMB           2004
#> 13 Sub-Saharan Africa         SSA       Zambia          ZMB           2006
#> 14 Sub-Saharan Africa         SSA       Zambia          ZMB           2010
#> 15 Sub-Saharan Africa         SSA       Zambia          ZMB           2015
#> 16 Sub-Saharan Africa         SSA       Zambia          ZMB           2022
#>    reporting_level survey_acronym survey_coverage survey_year welfare_type
#> 1         national           KIDS        national     1993.00  consumption
#> 2         national           HIES        national     2000.75  consumption
#> 3         national            IES        national     2005.00  consumption
#> 4         national            LCS        national     2008.67  consumption
#> 5         national            IES        national     2010.00  consumption
#> 6         national            LCS        national     2014.83  consumption
#> 7         national            HBS        national     1991.00  consumption
#> 8         national            HBS        national     1993.00  consumption
#> 9         national         LCMS-I        national     1996.00  consumption
#> 10        national        LCMS-II        national     1998.00  consumption
#> 11        national       LCMS-III        national     2002.83  consumption
#> 12        national        LCMS-IV        national     2004.67  consumption
#> 13        national         LCMS-V        national     2006.50  consumption
#> 14        national        LCMS-VI        national     2010.00  consumption
#> 15        national       LCMS-VII        national     2015.00  consumption
#> 16        national      LCMS-VIII        national     2022.00  consumption
#>    survey_comparability comparable_spell poverty_line headcount poverty_gap
#> 1                     0             1993         2.15 0.3354000  0.11971810
#> 2                     2             2000         2.15 0.3682500  0.14291195
#> 3                     3      2005 - 2014         2.15 0.2833133  0.09288520
#> 4                     3      2005 - 2014         2.15 0.1866297  0.05448886
#> 5                     3      2005 - 2014         2.15 0.1803596  0.05487818
#> 6                     3      2005 - 2014         2.15 0.2049256  0.06870356
#> 7                     0             1991         2.15 0.5743860  0.35951387
#> 8                     1             1993         2.15 0.5905600  0.30813174
#> 9                     2             1996         2.15 0.4677338  0.18942891
#> 10                    3      1998 - 2006         2.15 0.4722561  0.19783822
#> 11                    3      1998 - 2006         2.15 0.5638885  0.21314192
#> 12                    3      1998 - 2006         2.15 0.6106246  0.30677434
#> 13                    3      1998 - 2006         2.15 0.6469803  0.33600251
#> 14                    4      2010 - 2022         2.15 0.6440190  0.31418902
#> 15                    4      2010 - 2022         2.15 0.6079309  0.31802458
#> 16                    4      2010 - 2022         2.15 0.6434975  0.33094933
#>    poverty_severity      watts      mean   median       mld      gini
#> 1        0.05363426 0.16153411  7.512568 3.433418 0.6366259 0.5933394
#> 2        0.06945470 0.20017310  6.417391 3.118238 0.5999924 0.5776966
#> 3        0.04238680 0.12914374  9.539026 3.543823 0.7781977 0.6476209
#> 4        0.02290619 0.07313433 12.030882 4.754775 0.7351245 0.6300900
#> 5        0.02380089 0.07497618 12.693438 4.975511 0.7526825 0.6338269
#> 6        0.03252772 0.09708224 11.943078 4.753422 0.7502534 0.6302580
#> 7        0.26697867 0.72325000  3.119898 1.647333 0.7346184 0.5940293
#> 8        0.19953101 0.53839710  2.799716 1.706298 0.5083919 0.5255011
#> 9        0.09978449 0.28267047  3.619184 2.330480 0.4041692 0.4832838
#> 10       0.10749534 0.30277298  3.583934 2.277596 0.4229948 0.4912212
#> 11       0.10566047 0.30698124  2.732604 1.930888 0.2983276 0.4205863
#> 12       0.18804439 0.50586948  2.910987 1.609220 0.5216132 0.5426393
#> 13       0.21038845 0.56284589  2.680506 1.465410 0.5250066 0.5456934
#> 14       0.18728590 0.50608655  2.680243 1.537594 0.4672895 0.5201111
#> 15       0.20214089 0.54189873  2.976326 1.608135 0.5640261 0.5583758
#> 16       0.20631427 0.55236223  2.527800 1.485089 0.4655922 0.5148312
#>    polarization     decile1     decile2    decile3    decile4    decile5
#> 1     0.7147437 0.012500158 0.016969663 0.02260853 0.02990889 0.03968160
#> 2     0.6770340 0.012816826 0.017827628 0.02403983 0.03193381 0.04229243
#> 3     0.7549428 0.009987132 0.015865706 0.02091393 0.02623096 0.03312269
#> 4     0.7609605 0.010057582 0.016028558 0.02129058 0.02735578 0.03503969
#> 5     0.7837881 0.009383491 0.015307528 0.02053137 0.02663766 0.03442384
#> 6     0.7876801 0.008573526 0.015071925 0.02074484 0.02708844 0.03505582
#> 7     0.7582579 0.008024062 0.008667579 0.01582433 0.02725084 0.04317652
#> 8     0.5411877 0.011067811 0.019735176 0.02961919 0.04072260 0.05365048
#> 9     0.4541652 0.017246059 0.027708256 0.03687062 0.04620273 0.05726297
#> 10    0.4574639 0.016031158 0.027195332 0.03583848 0.04555774 0.05724019
#> 11    0.3480729 0.024213961 0.037077847 0.04643740 0.05537981 0.06517655
#> 12    0.5443317 0.013519311 0.022170250 0.03018672 0.03852603 0.04905841
#> 13    0.5484201 0.013640085 0.022110678 0.02962806 0.03823438 0.04860521
#> 14    0.5069726 0.015926239 0.024963175 0.03315291 0.04145005 0.05179524
#> 15    0.5838927 0.011593392 0.019786983 0.02742864 0.03583844 0.04743057
#> 16    0.5356241 0.014730372 0.023792915 0.03206024 0.04119664 0.05223906
#>       decile6    decile7    decile8   decile9  decile10          cpi      ppp
#> 1  0.05336964 0.07383648 0.10780574 0.1767219 0.4665975  0.250919243 6.549363
#> 2  0.05649406 0.07724597 0.11084411 0.1772000 0.4493053  0.427680670 6.549363
#> 3  0.04201139 0.05651415 0.08522598 0.1679379 0.5421901  0.523158205 6.549363
#> 4  0.04551465 0.06298532 0.09485078 0.1743129 0.5125641  0.643191146 6.549363
#> 5  0.04529468 0.06279671 0.09618910 0.1768473 0.5125883  0.717928132 6.549363
#> 6  0.04650066 0.06531552 0.09957491 0.1771186 0.5049558  0.893796432 6.549363
#> 7  0.06373679 0.09083097 0.12825572 0.1881510 0.4260822  0.002368799 4.223818
#> 8  0.06942920 0.09002134 0.11990938 0.1731933 0.3926515 17.831811990 4.223818
#> 9  0.07184293 0.08995160 0.11674991 0.1630785 0.3730865  0.053219872 4.223818
#> 10 0.07123715 0.08951033 0.11549372 0.1603932 0.3815027  0.082410771 4.223818
#> 11 0.07712225 0.09201815 0.11447193 0.1509535 0.3371486  0.230108468 4.223818
#> 12 0.06330786 0.08213626 0.11150286 0.1630353 0.4265570  0.314197420 4.223818
#> 13 0.06190788 0.08118061 0.11027127 0.1629910 0.4314308  0.380220828 4.223818
#> 14 0.06385541 0.08224747 0.11178960 0.1645366 0.4102833  0.552641656 4.223818
#> 15 0.06169146 0.08219847 0.11158199 0.1682373 0.4342127  0.796036473 4.223818
#> 16 0.06696812 0.08708625 0.11781478 0.1732533 0.3908583  1.839073644 4.223818
#>    reporting_pop reporting_gdp reporting_pce is_interpolated distribution_type
#> 1       42525440     4269.7002     2415.8990           FALSE             group
#> 2       47125602     4735.6655     2811.5279           FALSE             group
#> 3       49017147     5458.2302     3324.3577           FALSE             micro
#> 4       50971140     6074.9427     3779.0585           FALSE             micro
#> 5       51784921     6018.2308     3798.8744           FALSE             micro
#> 6       55681522     6252.3180     3948.9717           FALSE             micro
#> 7        7880466      867.7624            NA           FALSE             micro
#> 8        8270917      867.7130            NA           FALSE             group
#> 9        8902019      805.1419            NA           FALSE             micro
#> 10       9372430      790.8356            NA           FALSE             micro
#> 11      10781928      844.0920            NA           FALSE             micro
#> 12      11440516      907.4937            NA           FALSE             micro
#> 13      12186820      981.3455            NA           FALSE             micro
#> 14      13792086     1198.3048            NA           FALSE             micro
#> 15      16248230     1307.9096      662.4129           FALSE             micro
#> 16      20017675     1308.1018            NA           FALSE           imputed
#>    estimation_type      spl       spr        pg estimate_type
#> 1           survey 2.866709 0.4377100  9.782268            NA
#> 2           survey 2.709119 0.4506500 10.949629            NA
#> 3           survey 2.921912 0.4191221  9.251945            NA
#> 4           survey 3.527387 0.3796311  7.097350            NA
#> 5           survey 3.637755 0.3777062  7.060190            NA
#> 6           survey 3.526711 0.3856719  7.668721            NA
#> 7           survey 2.150000 0.5743860 32.953354            NA
#> 8           survey 2.150000 0.5905600 23.711825            NA
#> 9           survey 2.315240 0.4976326 14.467603            NA
#> 10          survey 2.288798 0.5024772 15.302274            NA
#> 11          survey 2.150000 0.5638885 15.518775            NA
#> 12          survey 2.150000 0.6106246 21.526754            NA
#> 13          survey 2.150000 0.6469803 23.388736            NA
#> 14          survey 2.150000 0.6440190 21.211359            NA
#> 15          survey 2.150000 0.6079309 23.087937            NA
#> 16          survey 2.150000 0.6434975 23.058921            NA
# }