Return main poverty and inequality statistics

Return main poverty and inequality statistics

Usage

pip_data(
  country = NULL,
  year = NULL,
  povline = 2.15,
  popshare = NULL,
  fill_gaps = FALSE,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  additional_ind = FALSE,
  release_version = NULL,
  ppp_version = NULL,
  version = NULL
)

Source

https://pip.worldbank.org/api

Arguments

country

(NULL | character())
Countries for which statistics are to be computed, specified as ISO3 codes. Default NULL.

year

(NULL | character() | numeric())
Years for which statistics are to be computed, specified as YYYY. Default NULL.

povline

(numeric(1))
Poverty line to be used to compute poverty mesures. Poverty lines are only accepted up to 3 decimals. Default 2.15.

popshare

(NULL | numeric(1))
Proportion of the population living below the poverty line. Will be ignored if povline is specified. Default NULL.

fill_gaps

(logical(1))
Whether to fill gaps in the data. Default FALSE.

welfare_type

(character(1))
Type of welfare measure to be used. Default "all".

reporting_level

(character(1))
level of reporting for the statistics. Default "all".

additional_ind

(logical(1))
Whether to include additional indicators. Default FALSE.

release_version

(NULL | character(1))
Version of the data release in YYYYMMDD format. Default NULL.

ppp_version

(NULL | character(1) | numeric(1))
Version of the data. Default NULL.

version

(NULL | character(1))
Version of the data. Default NULL.

Value

A data.frame() with the requested statistics.

See also

Examples

# \donttest{
pip_data(c("ZAF", "ZMB"))
#>           region_name region_code country_name country_code reporting_year
#> 1  Sub-Saharan Africa         SSA South Africa          ZAF           1993
#> 2  Sub-Saharan Africa         SSA South Africa          ZAF           2000
#> 3  Sub-Saharan Africa         SSA South Africa          ZAF           2005
#> 4  Sub-Saharan Africa         SSA South Africa          ZAF           2008
#> 5  Sub-Saharan Africa         SSA South Africa          ZAF           2010
#> 6  Sub-Saharan Africa         SSA South Africa          ZAF           2014
#> 7  Sub-Saharan Africa         SSA       Zambia          ZMB           1991
#> 8  Sub-Saharan Africa         SSA       Zambia          ZMB           1993
#> 9  Sub-Saharan Africa         SSA       Zambia          ZMB           1996
#> 10 Sub-Saharan Africa         SSA       Zambia          ZMB           1998
#> 11 Sub-Saharan Africa         SSA       Zambia          ZMB           2002
#> 12 Sub-Saharan Africa         SSA       Zambia          ZMB           2004
#> 13 Sub-Saharan Africa         SSA       Zambia          ZMB           2006
#> 14 Sub-Saharan Africa         SSA       Zambia          ZMB           2010
#> 15 Sub-Saharan Africa         SSA       Zambia          ZMB           2015
#> 16 Sub-Saharan Africa         SSA       Zambia          ZMB           2022
#>    reporting_level survey_acronym survey_coverage survey_year welfare_type
#> 1         national           KIDS        national     1993.00  consumption
#> 2         national           HIES        national     2000.75  consumption
#> 3         national            IES        national     2005.00  consumption
#> 4         national            LCS        national     2008.67  consumption
#> 5         national            IES        national     2010.00  consumption
#> 6         national            LCS        national     2014.83  consumption
#> 7         national            HBS        national     1991.00  consumption
#> 8         national            HBS        national     1993.00  consumption
#> 9         national         LCMS-I        national     1996.00  consumption
#> 10        national        LCMS-II        national     1998.00  consumption
#> 11        national       LCMS-III        national     2002.83  consumption
#> 12        national        LCMS-IV        national     2004.67  consumption
#> 13        national         LCMS-V        national     2006.50  consumption
#> 14        national        LCMS-VI        national     2010.00  consumption
#> 15        national       LCMS-VII        national     2015.00  consumption
#> 16        national      LCMS-VIII        national     2022.00  consumption
#>    survey_comparability comparable_spell poverty_line headcount poverty_gap
#> 1                     0             1993         2.15 0.3273500  0.11491448
#> 2                     2             2000         2.15 0.3602300  0.13788884
#> 3                     3      2005 - 2014         2.15 0.2738977  0.08866815
#> 4                     3      2005 - 2014         2.15 0.1784118  0.05158000
#> 5                     3      2005 - 2014         2.15 0.1736065  0.05210588
#> 6                     3      2005 - 2014         2.15 0.1983516  0.06567713
#> 7                     0             1991         2.15 0.5381684  0.33395130
#> 8                     1             1993         2.15 0.5424300  0.27433243
#> 9                     2             1996         2.15 0.4133771  0.15654076
#> 10                    3      1998 - 2006         2.15 0.4180117  0.16553755
#> 11                    3      1998 - 2006         2.15 0.4906109  0.17206749
#> 12                    3      1998 - 2006         2.15 0.5645888  0.27003818
#> 13                    3      1998 - 2006         2.15 0.6028530  0.29820089
#> 14                    4      2010 - 2022         2.15 0.5959820  0.27422965
#> 15                    4      2010 - 2022         2.15 0.5653299  0.28276612
#> 16                    4      2010 - 2022         2.15 0.5969237  0.29310627
#>    poverty_severity      watts      mean   median       mld      gini
#> 1        0.05071058 0.15409808  7.683054 3.511334 0.6366259 0.5933394
#> 2        0.06619755 0.19199955  6.563024 3.189002 0.5999924 0.5776966
#> 3        0.04015528 0.12285010  9.755512 3.624245 0.7781536 0.6476189
#> 4        0.02152328 0.06903565 12.303904 4.862677 0.7351245 0.6300900
#> 5        0.02242921 0.07096926 12.981507 5.088423 0.7526440 0.6338256
#> 6        0.03093106 0.09252674 12.214114 4.861294 0.7502294 0.6302572
#> 7        0.24502732 0.65643750  3.527235 1.862636 0.7360095 0.5942023
#> 8        0.17391152 0.46908923  3.165553 1.929307 0.5086483 0.5255385
#> 9        0.07927822 0.22857635  4.092197 2.635068 0.4041857 0.4832856
#> 10       0.08672670 0.24822724  4.052331 2.575272 0.4230418 0.4912264
#> 11       0.08129958 0.24229162  3.089747 2.183250 0.2983345 0.4205873
#> 12       0.15994412 0.43382644  3.291411 1.819541 0.5217250 0.5426552
#> 13       0.18051083 0.48622575  3.030802 1.656936 0.5251344 0.5457134
#> 14       0.15732420 0.43006570  3.030522 1.738554 0.4673595 0.5201220
#> 15       0.17465904 0.46996718  3.365255 1.818314 0.5642529 0.5584075
#> 16       0.17671518 0.47631045  2.858142 1.679187 0.4657032 0.5148495
#>    polarization     decile1    decile2    decile3    decile4    decile5
#> 1     0.7147437 0.012500158 0.01696966 0.02260853 0.02990889 0.03968160
#> 2     0.6770340 0.012816826 0.01782763 0.02403983 0.03193381 0.04229243
#> 3     0.7549428 0.009988367 0.01586569 0.02091390 0.02623093 0.03312265
#> 4     0.7609605 0.010057582 0.01602856 0.02129058 0.02735578 0.03503969
#> 5     0.7837881 0.009384293 0.01530752 0.02053135 0.02663764 0.03442381
#> 6     0.7876801 0.008574058 0.01507192 0.02074483 0.02708842 0.03505580
#> 7     0.7583306 0.007949105 0.00862397 0.01582624 0.02725412 0.04318172
#> 8     0.5411901 0.011043074 0.01973567 0.02961993 0.04072362 0.05365183
#> 9     0.4541652 0.017244862 0.02770829 0.03687066 0.04620279 0.05726304
#> 10    0.4574639 0.016027705 0.02719543 0.03583861 0.04555790 0.05724039
#> 11    0.3480729 0.024213292 0.03707787 0.04643743 0.05537985 0.06517659
#> 12    0.5443322 0.013508999 0.02217048 0.03018703 0.03852643 0.04905893
#> 13    0.5484208 0.013627192 0.02211097 0.02962845 0.03823488 0.04860585
#> 14    0.5069728 0.015919193 0.02496335 0.03315314 0.04145034 0.05179562
#> 15    0.5838948 0.011572949 0.01978739 0.02742921 0.03583918 0.04743155
#> 16    0.5356247 0.014718426 0.02379320 0.03206063 0.04119714 0.05223970
#>       decile6    decile7    decile8   decile9  decile10          cpi      ppp
#> 1  0.05336964 0.07383648 0.10780574 0.1767219 0.4665975  0.213554679 7.524514
#> 2  0.05649406 0.07724597 0.11084411 0.1772000 0.4493053  0.363994436 7.524514
#> 3  0.04201133 0.05651407 0.08522587 0.1679377 0.5421895  0.445254343 7.524514
#> 4  0.04551465 0.06298532 0.09485078 0.1743129 0.5125641  0.547413093 7.524514
#> 5  0.04529464 0.06279666 0.09618902 0.1768472 0.5125879  0.611020941 7.524514
#> 6  0.04650064 0.06531549 0.09957486 0.1771185 0.5049555  0.760700567 7.524514
#> 7  0.06374448 0.09084192 0.12827118 0.1881736 0.4261336  0.001429636 6.189585
#> 8  0.06943093 0.09002360 0.11991238 0.1731977 0.3926613 10.761994155 6.189585
#> 9  0.07184301 0.08995171 0.11675005 0.1630787 0.3730869  0.032119672 6.189585
#> 10 0.07123740 0.08951064 0.11549412 0.1603937 0.3815041  0.049737191 6.189585
#> 11 0.07712230 0.09201821 0.11447201 0.1509536 0.3371489  0.138876856 6.189585
#> 12 0.06330852 0.08213712 0.11150403 0.1630370 0.4265615  0.189626876 6.189585
#> 13 0.06190869 0.08118167 0.11027271 0.1629932 0.4314364  0.229473838 6.189585
#> 14 0.06385587 0.08224806 0.11179040 0.1645378 0.4102862  0.333534600 6.189585
#> 15 0.06169273 0.08220017 0.11158430 0.1682408 0.4342217  0.480430136 6.189585
#> 16 0.06696894 0.08708731 0.11781621 0.1732554 0.3908631  1.109932059 6.189585
#>    reporting_pop reporting_gdp reporting_pce is_interpolated distribution_type
#> 1       43297156     4193.5983     2372.8387           FALSE             group
#> 2       47465030     4700.8756     2790.8733           FALSE             group
#> 3       49490033     5406.0758     3292.5929           FALSE             micro
#> 4       51525923     6009.7196     3738.4850           FALSE             micro
#> 5       52344051     5953.9451     3758.2954           FALSE             micro
#> 6       56531658     6155.0059     3887.5093           FALSE             micro
#> 7        7981650      856.7617            NA           FALSE             micro
#> 8        8373921      857.0397            NA           FALSE             group
#> 9        9004053      796.0181            NA           FALSE             micro
#> 10       9482408      781.6634            NA           FALSE             micro
#> 11      10926535      833.0212            NA           FALSE             micro
#> 12      11593214      895.4753            NA           FALSE             micro
#> 13      12347319      968.5636            NA           FALSE             micro
#> 14      13965594     1183.4171            NA           FALSE             micro
#> 15      16399089     1295.8779      603.8121           FALSE             micro
#> 16      20152938     1298.8480            NA           FALSE           imputed
#>    estimation_type   spl       spr        pg estimate_type
#> 1           survey 3.056 0.4522100 10.713024            NA
#> 2           survey 3.000 0.4786800 11.991457            NA
#> 3           survey 3.112 0.4359328 10.127501            NA
#> 4           survey 3.731 0.3929107  7.772644            NA
#> 5           survey 3.844 0.3938811  7.727838            NA
#> 6           survey 3.731 0.3995142  8.395812            NA
#> 7           survey 3.000 0.6298410 32.792041            NA
#> 8           survey 3.000 0.6705900 23.515498            NA
#> 9           survey 3.000 0.5606018 14.332490            NA
#> 10          survey 3.000 0.5662368 15.162530            NA
#> 11          survey 3.000 0.6798157 15.372702            NA
#> 12          survey 3.000 0.6893102 21.335260            NA
#> 13          survey 3.000 0.7196303 23.181471            NA
#> 14          survey 3.000 0.7136719 21.018355            NA
#> 15          survey 3.000 0.6789284 22.894125            NA
#> 16          survey 3.000 0.7165615 22.853020            NA
# }