Return aggregation of PIP statistics

Return aggregation of PIP statistics

Usage

pip_group(
  country = NULL,
  year = NULL,
  povline = 2.15,
  popshare = NULL,
  group_by = c("wb", "none"),
  fill_gaps = FALSE,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  additional_ind = FALSE,
  release_version = NULL,
  ppp_version = NULL,
  version = NULL
)

Source

https://pip.worldbank.org/api

Arguments

country

(character()) countries for which statistics are to be computed, specified as ISO3 codes. Default NULL.

year

(character() | numeric()) year(s) for which statistics are to be computed, specified as YYYY. Default NULL.

povline

(numeric(1)) poverty line to be used to compute poverty mesures. Poverty lines are only accepted up to 3 decimals. Default 2.15.

popshare

(numeric(1)) proportion of the population living below the poverty line. Will be ignored if povline is specified. Default NULL.

group_by

(character(1)) aggregate results by pre-defined sub-groups. Default "wb".

fill_gaps

(logical(1)) whether to fill gaps in the data. Default FALSE.

welfare_type

(character(1)) type of welfare measure to be used. Default "all".

reporting_level

(character(1)) level of reporting for the statistics. Default "all".

additional_ind

(logical(1)) whether to include additional indicators. Default FALSE.

release_version

(character(1)) version of the data release in YYYYMMDD format. Default NULL.

ppp_version

(character(1) | numeric(1)) version of the data. Default NULL.

version

(character(1)) version of the data. Default NULL.

Value

A data.frame() with the requested statistics.

See also

Examples

# \donttest{
pip_group(c("AFE", "LAC"))
#>                    region_name region_code reporting_year poverty_line
#> 1  Eastern and Southern Africa         AFE           1981         2.15
#> 2  Eastern and Southern Africa         AFE           1982         2.15
#> 3  Eastern and Southern Africa         AFE           1983         2.15
#> 4  Eastern and Southern Africa         AFE           1984         2.15
#> 5  Eastern and Southern Africa         AFE           1985         2.15
#> 6  Eastern and Southern Africa         AFE           1986         2.15
#> 7  Eastern and Southern Africa         AFE           1987         2.15
#> 8  Eastern and Southern Africa         AFE           1988         2.15
#> 9  Eastern and Southern Africa         AFE           1989         2.15
#> 10 Eastern and Southern Africa         AFE           1990         2.15
#> 11 Eastern and Southern Africa         AFE           1991         2.15
#> 12 Eastern and Southern Africa         AFE           1992         2.15
#> 13 Eastern and Southern Africa         AFE           1993         2.15
#> 14 Eastern and Southern Africa         AFE           1994         2.15
#> 15 Eastern and Southern Africa         AFE           1995         2.15
#> 16 Eastern and Southern Africa         AFE           1996         2.15
#> 17 Eastern and Southern Africa         AFE           1997         2.15
#> 18 Eastern and Southern Africa         AFE           1998         2.15
#> 19 Eastern and Southern Africa         AFE           1999         2.15
#> 20 Eastern and Southern Africa         AFE           2000         2.15
#> 21 Eastern and Southern Africa         AFE           2001         2.15
#> 22 Eastern and Southern Africa         AFE           2002         2.15
#> 23 Eastern and Southern Africa         AFE           2003         2.15
#> 24 Eastern and Southern Africa         AFE           2004         2.15
#> 25 Eastern and Southern Africa         AFE           2005         2.15
#> 26 Eastern and Southern Africa         AFE           2006         2.15
#> 27 Eastern and Southern Africa         AFE           2007         2.15
#> 28 Eastern and Southern Africa         AFE           2008         2.15
#> 29 Eastern and Southern Africa         AFE           2009         2.15
#> 30 Eastern and Southern Africa         AFE           2010         2.15
#> 31 Eastern and Southern Africa         AFE           2011         2.15
#> 32 Eastern and Southern Africa         AFE           2012         2.15
#> 33 Eastern and Southern Africa         AFE           2013         2.15
#> 34 Eastern and Southern Africa         AFE           2014         2.15
#> 35 Eastern and Southern Africa         AFE           2015         2.15
#> 36 Eastern and Southern Africa         AFE           2016         2.15
#> 37 Eastern and Southern Africa         AFE           2017         2.15
#> 38 Eastern and Southern Africa         AFE           2018         2.15
#> 39 Eastern and Southern Africa         AFE           2019         2.15
#> 40 Eastern and Southern Africa         AFE           2020         2.15
#> 41 Eastern and Southern Africa         AFE           2021         2.15
#> 42 Eastern and Southern Africa         AFE           2022         2.15
#> 43 Eastern and Southern Africa         AFE           2023         2.15
#> 44 Eastern and Southern Africa         AFE           2024         2.15
#> 45 Eastern and Southern Africa         AFE           2025         2.15
#> 46   Latin America & Caribbean         LAC           1981         2.15
#> 47   Latin America & Caribbean         LAC           1982         2.15
#> 48   Latin America & Caribbean         LAC           1983         2.15
#> 49   Latin America & Caribbean         LAC           1984         2.15
#> 50   Latin America & Caribbean         LAC           1985         2.15
#> 51   Latin America & Caribbean         LAC           1986         2.15
#> 52   Latin America & Caribbean         LAC           1987         2.15
#> 53   Latin America & Caribbean         LAC           1988         2.15
#> 54   Latin America & Caribbean         LAC           1989         2.15
#> 55   Latin America & Caribbean         LAC           1990         2.15
#> 56   Latin America & Caribbean         LAC           1991         2.15
#> 57   Latin America & Caribbean         LAC           1992         2.15
#> 58   Latin America & Caribbean         LAC           1993         2.15
#> 59   Latin America & Caribbean         LAC           1994         2.15
#> 60   Latin America & Caribbean         LAC           1995         2.15
#> 61   Latin America & Caribbean         LAC           1996         2.15
#> 62   Latin America & Caribbean         LAC           1997         2.15
#> 63   Latin America & Caribbean         LAC           1998         2.15
#> 64   Latin America & Caribbean         LAC           1999         2.15
#> 65   Latin America & Caribbean         LAC           2000         2.15
#> 66   Latin America & Caribbean         LAC           2001         2.15
#> 67   Latin America & Caribbean         LAC           2002         2.15
#> 68   Latin America & Caribbean         LAC           2003         2.15
#> 69   Latin America & Caribbean         LAC           2004         2.15
#> 70   Latin America & Caribbean         LAC           2005         2.15
#> 71   Latin America & Caribbean         LAC           2006         2.15
#> 72   Latin America & Caribbean         LAC           2007         2.15
#> 73   Latin America & Caribbean         LAC           2008         2.15
#> 74   Latin America & Caribbean         LAC           2009         2.15
#> 75   Latin America & Caribbean         LAC           2010         2.15
#> 76   Latin America & Caribbean         LAC           2011         2.15
#> 77   Latin America & Caribbean         LAC           2012         2.15
#> 78   Latin America & Caribbean         LAC           2013         2.15
#> 79   Latin America & Caribbean         LAC           2014         2.15
#> 80   Latin America & Caribbean         LAC           2015         2.15
#> 81   Latin America & Caribbean         LAC           2016         2.15
#> 82   Latin America & Caribbean         LAC           2017         2.15
#> 83   Latin America & Caribbean         LAC           2018         2.15
#> 84   Latin America & Caribbean         LAC           2019         2.15
#> 85   Latin America & Caribbean         LAC           2020         2.15
#> 86   Latin America & Caribbean         LAC           2021         2.15
#> 87   Latin America & Caribbean         LAC           2022         2.15
#> 88   Latin America & Caribbean         LAC           2023         2.15
#> 89   Latin America & Caribbean         LAC           2024         2.15
#> 90   Latin America & Caribbean         LAC           2025         2.15
#>     headcount poverty_gap poverty_severity      watts      mean       spr
#> 1  0.43340749 0.171934034      0.090856935 0.25858601  4.749965 0.6065094
#> 2  0.43821253 0.175573474      0.093475701 0.26543299  4.686368 0.6093514
#> 3  0.44491739 0.180465027      0.097131387 0.27517446  4.607832 0.6137755
#> 4  0.45421984 0.185406557      0.100150094 0.28362496  4.551052 0.6218635
#> 5  0.46282488 0.189705530      0.102655394 0.29039398  4.481283 0.6277353
#> 6  0.45620201 0.186632449      0.101076541 0.28613490  4.506459 0.6229176
#> 7  0.45091152 0.183065780      0.098495012 0.27904157  4.559732 0.6182811
#> 8  0.44949796 0.181434659      0.097070203 0.27528249  4.601891 0.6184546
#> 9  0.45084347 0.182404642      0.097605302 0.27657828  4.608459 0.6201085
#> 10 0.46030225 0.187431108      0.100769334 0.28490919  4.548679 0.6272347
#> 11 0.47725146 0.198304399      0.107673331 0.30291437  4.456959 0.6392894
#> 12 0.49222922 0.207486276      0.113634952 0.31847846  4.319305 0.6501096
#> 13 0.50421770 0.215507932      0.119327710 0.33282487  4.050096 0.6568586
#> 14 0.50841273 0.222261773      0.125106309 0.34708586  3.964155 0.6605470
#> 15 0.50058166 0.217145389      0.121896335 0.33861137  4.046432 0.6551652
#> 16 0.48934248 0.211607474      0.118658808 0.32966323  4.112409 0.6489526
#> 17 0.48774608 0.212263936      0.119765844 0.33217439  4.116267 0.6495245
#> 18 0.48828117 0.213055162      0.120643164 0.33428381  4.053095 0.6524412
#> 19 0.48510750 0.212588971      0.121113767 0.33528398  3.975389 0.6522526
#> 20 0.48203845 0.215249754      0.124400862 0.34324609  3.928387 0.6489951
#> 21 0.47953415 0.214276793      0.124146527 0.34270808  3.915873 0.6448766
#> 22 0.48529411 0.212043596      0.121339646 0.33701591  3.930173 0.6524805
#> 23 0.47607489 0.204438386      0.115935411 0.32320136  3.965661 0.6438843
#> 24 0.44224748 0.189884646      0.108168519 0.30086587  4.206457 0.6094818
#> 25 0.42151876 0.180567241      0.102649740 0.28540499  4.415301 0.5931298
#> 26 0.40240367 0.169469349      0.094929432 0.26489940  4.671371 0.5808191
#> 27 0.38948954 0.160366974      0.087930219 0.24679614  4.743310 0.5712187
#> 28 0.37838379 0.152126475      0.081796896 0.23111185  4.828392 0.5631981
#> 29 0.37593498 0.151931847      0.082105920 0.23178233  4.815238 0.5601640
#> 30 0.36253041 0.143402674      0.076115331 0.21629141  4.946660 0.5496669
#> 31 0.35132651 0.136842404      0.071966707 0.20545487  5.023404 0.5405087
#> 32 0.34322170 0.131694622      0.068522091 0.19671527  5.037383 0.5359166
#> 33 0.33864194 0.128550324      0.066538588 0.19152274  5.054164 0.5315254
#> 34 0.33776004 0.128203128      0.066304002 0.19089153  5.062764 0.5295755
#> 35 0.34086947 0.129860176      0.067280140 0.19351648  5.045060 0.5312359
#> 36 0.34872619 0.134136291      0.069879552 0.20051865  4.987404 0.5374672
#> 37 0.35705284 0.137868079      0.072171874 0.20668248  4.903704 0.5422655
#> 38 0.36190938 0.140180808      0.073663666 0.21059672  4.838839 0.5457029
#> 39 0.36681789 0.142986516      0.075539118 0.21544714  4.753060 0.5487115
#> 40 0.38170585 0.151567138      0.081150500 0.23025775  4.529296 0.5610098
#> 41 0.37867799 0.148938052      0.079414591 0.22571876  4.539248 0.5613019
#> 42 0.37211527 0.145645223      0.077307330 0.22009038  4.574368 0.5567246
#> 43 0.36761984 0.142834949      0.075411364 0.21513504  4.559430 0.5540615
#> 44 0.36875900 0.143565498      0.076030838 0.21675154  4.533114 0.5537737
#> 45 0.36307844 0.141434976      0.075002113 0.21371013  4.576789 0.5483001
#> 46 0.12758111 0.049459160      0.027767321 0.07971246 12.685808 0.3481803
#> 47 0.13347845 0.051361425      0.028699650 0.08247968 12.564294 0.3517465
#> 48 0.15872419 0.061639716      0.034206873 0.09834164 11.729736 0.3648431
#> 49 0.15511219 0.059534865      0.033043340 0.09519231 11.816001 0.3621156
#> 50 0.14200873 0.054793986      0.030590110 0.08810016 11.870391 0.3572308
#> 51 0.10084309 0.039090730      0.022569992 0.06481986 14.477099 0.3359948
#> 52 0.12486029 0.049421178      0.028333168 0.08103435 13.143611 0.3480899
#> 53 0.13173340 0.054270677      0.031846061 0.09027199 12.896922 0.3478761
#> 54 0.12725453 0.052807886      0.031145174 0.08831886 13.138433 0.3451173
#> 55 0.13822575 0.055769636      0.031736232 0.09041367 11.991365 0.3500661
#> 56 0.13265488 0.053644369      0.030939551 0.08795731 11.814562 0.3456009
#> 57 0.13244217 0.054638398      0.032176633 0.09103167 11.460216 0.3451453
#> 58 0.13245127 0.054322383      0.031909952 0.09050038 12.228710 0.3503307
#> 59 0.12736623 0.052118446      0.030683918 0.08734684 12.623060 0.3487113
#> 60 0.11795163 0.047767824      0.027942746 0.07981077 13.117059 0.3442966
#> 61 0.13025346 0.055149877      0.033076979 0.09347332 12.952346 0.3520986
#> 62 0.12907547 0.054294911      0.032318246 0.09130682 13.258961 0.3519357
#> 63 0.12751906 0.053697954      0.031891261 0.09024368 13.528425 0.3488033
#> 64 0.12977511 0.054683928      0.032694675 0.09254658 13.076648 0.3511054
#> 65 0.11849733 0.050652825      0.030758173 0.08677293 13.727282 0.3451592
#> 66 0.11580028 0.048188424      0.028300881 0.07999753 13.778011 0.3435557
#> 67 0.10702005 0.041729366      0.023337784 0.06691386 13.615806 0.3418449
#> 68 0.10377128 0.040878110      0.023128326 0.06606511 13.703200 0.3367531
#> 69 0.09455180 0.036613282      0.020472345 0.05872706 13.974827 0.3301083
#> 70 0.09041813 0.035731697      0.020271081 0.05777076 14.768119 0.3258097
#> 71 0.07459130 0.028837648      0.016080192 0.04603129 15.851153 0.3147439
#> 72 0.06347560 0.023555592      0.012839418 0.03714898 16.446853 0.3086224
#> 73 0.05739372 0.021046851      0.011339549 0.03297398 16.995347 0.3023136
#> 74 0.05495404 0.019698157      0.010455045 0.03062429 16.993047 0.2985083
#> 75 0.04853226 0.017481788      0.009306949 0.02725444 17.582610 0.2909722
#> 76 0.04425843 0.016215790      0.008704368 0.02540209 18.312843 0.2852065
#> 77 0.04116925 0.014668458      0.007687096 0.02268364 19.032441 0.2845864
#> 78 0.03672052 0.012500528      0.006497476 0.01933329 19.522933 0.2793164
#> 79 0.03426283 0.011415129      0.005961698 0.01782960 19.527259 0.2751645
#> 80 0.03166772 0.010249389      0.005080003 0.01541882 19.937569 0.2722115
#> 81 0.03424041 0.011500410      0.005791601 0.01733837 20.110826 0.2739847
#> 82 0.03482042 0.011790789      0.005919324 0.01776856 20.262932 0.2710610
#> 83 0.03432246 0.011716158      0.005899791 0.01763223 20.456772 0.2692400
#> 84 0.03510453 0.012025347      0.006128680 0.01825649 20.495733 0.2715972
#> 85 0.03153234 0.011170711      0.006028925 0.01770432 19.391203 0.2576947
#> 86 0.03679714 0.012645172      0.006375547 0.01902979 19.584810 0.2687202
#> 87 0.02842836 0.009863431      0.005134792 0.01526176 20.730196 0.2588591
#> 88 0.02508914 0.008866915      0.004764375 0.01401903 21.393340 0.2587329
#> 89 0.02471605 0.008811457      0.004744323 0.01392815 21.706600 0.2582089
#> 90 0.02454394 0.008746876      0.004711686 0.01381392 21.979109 0.2598353
#>           pg estimate_type reporting_pop pop_in_poverty
#> 1  15.325681    projection     238043099      103169662
#> 2  15.584901    projection     245822010      107722286
#> 3  15.950515    projection     253644643      112850912
#> 4  16.255413    projection     261458202      118759502
#> 5  16.485747    projection     269450407      124708352
#> 6  16.367415    projection     277621771      126651611
#> 7  16.095038    projection     286067346      128991062
#> 8  15.940664    projection     294498625      132376532
#> 9  15.964918    projection     302939121      136578125
#> 10 16.253345    projection     311748681      143498618
#> 11 16.820852    projection     320442961      152931872
#> 12 17.330840    projection     329082707      161984126
#> 13 17.856235        actual     338324002      170588949
#> 14 18.358529        actual     347441809      176643840
#> 15 18.078346        actual     356580375      178497597
#> 16 17.790233        actual     366138524      179167135
#> 17 17.905841    projection     375646235      183219977
#> 18 18.020417        actual     385505757      188235204
#> 19 18.153056        actual     395750933      191981746
#> 20 18.504775        actual     406156661      195783128
#> 21 18.500903        actual     416807868      199873606
#> 22 18.352193        actual     427820358      207618700
#> 23 17.918977        actual     439173286      209079373
#> 24 17.140157        actual     450928044      199421791
#> 25 16.559836        actual     463076637      195195490
#> 26 15.719029        actual     475606210      191385683
#> 27 15.008785        actual     488580707      190297076
#> 28 14.427666        actual     502070763      189975437
#> 29 14.466325        actual     516003448      193983748
#> 30 13.915081        actual     530308387      192252918
#> 31 13.542848        actual     544737983      191380896
#> 32 13.277582        actual     559609961      192070281
#> 33 13.104822        actual     575202699      194787758
#> 34 13.066818        actual     590968990      199605709
#> 35 13.141451        actual     607123269      206949786
#> 36 13.377883        actual     623369401      217385237
#> 37 13.616662        actual     640058741      228534788
#> 38 13.783042        actual     657801085      238064380
#> 39 13.983763    projection     675950189      247950619
#> 40 14.563503        actual     694446100      265074138
#> 41 14.454736        actual     713090928      270031839
#> 42 14.244015        actual     731821393      272321919
#> 43 14.092019    projection     750503764      275900073
#> 44 14.164284       nowcast     769294967      283684442
#> 45 14.034081       nowcast     788858910      286417661
#> 46  7.497425        actual     363707276       46402177
#> 47  7.623677        actual     371611051       49602067
#> 48  8.401517        actual     379654360       60260331
#> 49  8.305235        actual     387785677       60150284
#> 50  7.954611        actual     395922883       56224508
#> 51  6.745903        actual     404042411       40744886
#> 52  7.529395        actual     412162093       51462680
#> 53  7.891745        actual     420314572       55369469
#> 54  7.818285        actual     428474147       54525277
#> 55  7.929936        actual     436660341       60357703
#> 56  7.839854        actual     444793033       59003965
#> 57  8.037451        actual     452872282       59979387
#> 58  7.986493        actual     460906825       61047696
#> 59  7.785102        actual     468862052       59717190
#> 60  7.519271        actual     476796636       56238940
#> 61  8.084502        actual     484667880       63129669
#> 62  7.901111        actual     492482495       63567412
#> 63  7.830460        actual     500308570       63798876
#> 64  7.994837        actual     508063703       65934021
#> 65  7.658184        actual     515590262       61096069
#> 66  7.299803        actual     522877879       60549405
#> 67  6.792013        actual     529982351       56718738
#> 68  6.716480        actual     536918721       55716743
#> 69  6.349273        actual     543716438       51409369
#> 70  6.155342        actual     550371741       49763583
#> 71  5.479650        actual     557003859       41547644
#> 72  5.034154        actual     563605218       35775180
#> 73  4.777429        actual     570177672       32724619
#> 74  4.658420        actual     576665478       31690096
#> 75  4.398511        actual     583034181       28295965
#> 76  4.218144        actual     589387094       26085350
#> 77  4.052175        actual     595740249       24526180
#> 78  3.859222        actual     602020041       22106488
#> 79  3.795360        actual     608220429       20839353
#> 80  3.600245        actual     614277357       19452763
#> 81  3.654151        actual     620192706       21235652
#> 82  3.624599        actual     625934025       21795285
#> 83  3.596558        actual     631410622       21671569
#> 84  3.604058        actual     636578789       22346802
#> 85  3.641482        actual     641135529       20216504
#> 86  3.678749        actual     644892429       23730196
#> 87  3.390942        actual     648804795       18444458
#> 88  3.269550        actual     653324554       16391352
#> 89  3.203427       nowcast     657893350       16260528
#> 90  3.174818       nowcast     662317635       16255883
# }