Return aggregation of PIP statistics

Return aggregation of PIP statistics

Usage

pip_group(
  country = NULL,
  year = NULL,
  povline = 2.15,
  popshare = NULL,
  group_by = c("wb", "none"),
  fill_gaps = FALSE,
  welfare_type = c("all", "consumption", "income"),
  reporting_level = c("all", "national", "rural", "urban"),
  additional_ind = FALSE,
  release_version = NULL,
  ppp_version = NULL,
  version = NULL
)

Source

https://pip.worldbank.org/api

Arguments

country

(NULL | character())
Countries for which statistics are to be computed, specified as ISO3 codes. Default NULL.

year

(NULL | character() | numeric())
Years for which statistics are to be computed, specified as YYYY. Default NULL.

povline

(numeric(1))
Poverty line to be used to compute poverty mesures. Poverty lines are only accepted up to 3 decimals. Default 2.15.

popshare

(NULL | numeric(1))
Proportion of the population living below the poverty line. Will be ignored if povline is specified. Default NULL.

group_by

(character(1))
Aggregate results by pre-defined sub-groups. Default "wb".

fill_gaps

(logical(1))
Whether to fill gaps in the data. Default FALSE.

welfare_type

(character(1))
Type of welfare measure to be used. Default "all".

reporting_level

(character(1))
level of reporting for the statistics. Default "all".

additional_ind

(logical(1))
Whether to include additional indicators. Default FALSE.

release_version

(NULL | character(1))
Version of the data release in YYYYMMDD format. Default NULL.

ppp_version

(NULL | character(1) | numeric(1))
Version of the data. Default NULL.

version

(NULL | character(1))
Version of the data. Default NULL.

Value

A data.frame() with the requested statistics.

See also

Examples

# \donttest{
pip_group(c("AFE", "LAC"))
#>                    region_name region_code reporting_year reporting_pop
#> 1  Eastern and Southern Africa         AFE           1981     238043099
#> 2  Eastern and Southern Africa         AFE           1982     245822010
#> 3  Eastern and Southern Africa         AFE           1983     253644643
#> 4  Eastern and Southern Africa         AFE           1984     261458202
#> 5  Eastern and Southern Africa         AFE           1985     269450407
#> 6  Eastern and Southern Africa         AFE           1986     277621771
#> 7  Eastern and Southern Africa         AFE           1987     286067346
#> 8  Eastern and Southern Africa         AFE           1988     294498625
#> 9  Eastern and Southern Africa         AFE           1989     302939121
#> 10 Eastern and Southern Africa         AFE           1990     311748681
#> 11 Eastern and Southern Africa         AFE           1991     320442961
#> 12 Eastern and Southern Africa         AFE           1992     329082707
#> 13 Eastern and Southern Africa         AFE           1993     338324002
#> 14 Eastern and Southern Africa         AFE           1994     347441809
#> 15 Eastern and Southern Africa         AFE           1995     356580375
#> 16 Eastern and Southern Africa         AFE           1996     366138524
#> 17 Eastern and Southern Africa         AFE           1997     375646235
#> 18 Eastern and Southern Africa         AFE           1998     385505757
#> 19 Eastern and Southern Africa         AFE           1999     395750933
#> 20 Eastern and Southern Africa         AFE           2000     406156661
#> 21 Eastern and Southern Africa         AFE           2001     416807868
#> 22 Eastern and Southern Africa         AFE           2002     427820358
#> 23 Eastern and Southern Africa         AFE           2003     439173286
#> 24 Eastern and Southern Africa         AFE           2004     450928044
#> 25 Eastern and Southern Africa         AFE           2005     463076637
#> 26 Eastern and Southern Africa         AFE           2006     475606210
#> 27 Eastern and Southern Africa         AFE           2007     488580707
#> 28 Eastern and Southern Africa         AFE           2008     502070763
#> 29 Eastern and Southern Africa         AFE           2009     516003448
#> 30 Eastern and Southern Africa         AFE           2010     530308387
#> 31 Eastern and Southern Africa         AFE           2011     544737983
#> 32 Eastern and Southern Africa         AFE           2012     559609961
#> 33 Eastern and Southern Africa         AFE           2013     575202699
#> 34 Eastern and Southern Africa         AFE           2014     590968990
#> 35 Eastern and Southern Africa         AFE           2015     607123269
#> 36 Eastern and Southern Africa         AFE           2016     623369401
#> 37 Eastern and Southern Africa         AFE           2017     640058741
#> 38 Eastern and Southern Africa         AFE           2018     657801085
#> 39 Eastern and Southern Africa         AFE           2019     675950189
#> 40 Eastern and Southern Africa         AFE           2020     694446100
#> 41 Eastern and Southern Africa         AFE           2021     713090928
#> 42 Eastern and Southern Africa         AFE           2022     731821393
#> 43 Eastern and Southern Africa         AFE           2023     750503764
#> 44 Eastern and Southern Africa         AFE           2024     769294967
#> 45 Eastern and Southern Africa         AFE           2025     788858910
#> 46   Latin America & Caribbean         LAC           1981     363707276
#> 47   Latin America & Caribbean         LAC           1982     371611051
#> 48   Latin America & Caribbean         LAC           1983     379654360
#> 49   Latin America & Caribbean         LAC           1984     387785677
#> 50   Latin America & Caribbean         LAC           1985     395922883
#> 51   Latin America & Caribbean         LAC           1986     404042411
#> 52   Latin America & Caribbean         LAC           1987     412162093
#> 53   Latin America & Caribbean         LAC           1988     420314572
#> 54   Latin America & Caribbean         LAC           1989     428474147
#> 55   Latin America & Caribbean         LAC           1990     436660341
#> 56   Latin America & Caribbean         LAC           1991     444793033
#> 57   Latin America & Caribbean         LAC           1992     452872282
#> 58   Latin America & Caribbean         LAC           1993     460906825
#> 59   Latin America & Caribbean         LAC           1994     468862052
#> 60   Latin America & Caribbean         LAC           1995     476796636
#> 61   Latin America & Caribbean         LAC           1996     484667880
#> 62   Latin America & Caribbean         LAC           1997     492482495
#> 63   Latin America & Caribbean         LAC           1998     500308570
#> 64   Latin America & Caribbean         LAC           1999     508063703
#> 65   Latin America & Caribbean         LAC           2000     515590262
#> 66   Latin America & Caribbean         LAC           2001     522877879
#> 67   Latin America & Caribbean         LAC           2002     529982351
#> 68   Latin America & Caribbean         LAC           2003     536918721
#> 69   Latin America & Caribbean         LAC           2004     543716438
#> 70   Latin America & Caribbean         LAC           2005     550371741
#> 71   Latin America & Caribbean         LAC           2006     557003859
#> 72   Latin America & Caribbean         LAC           2007     563605218
#> 73   Latin America & Caribbean         LAC           2008     570177672
#> 74   Latin America & Caribbean         LAC           2009     576665478
#> 75   Latin America & Caribbean         LAC           2010     583034181
#> 76   Latin America & Caribbean         LAC           2011     589387094
#> 77   Latin America & Caribbean         LAC           2012     595740249
#> 78   Latin America & Caribbean         LAC           2013     602020041
#> 79   Latin America & Caribbean         LAC           2014     608220429
#> 80   Latin America & Caribbean         LAC           2015     614277357
#> 81   Latin America & Caribbean         LAC           2016     620192706
#> 82   Latin America & Caribbean         LAC           2017     625934025
#> 83   Latin America & Caribbean         LAC           2018     631410622
#> 84   Latin America & Caribbean         LAC           2019     636578789
#> 85   Latin America & Caribbean         LAC           2020     641135529
#> 86   Latin America & Caribbean         LAC           2021     644892429
#> 87   Latin America & Caribbean         LAC           2022     648804795
#> 88   Latin America & Caribbean         LAC           2023     653324554
#> 89   Latin America & Caribbean         LAC           2024     657893350
#> 90   Latin America & Caribbean         LAC           2025     662317635
#>    poverty_line  headcount poverty_gap poverty_severity      watts      mean
#> 1          2.15 0.43340749 0.171934034      0.090856935 0.25858601  4.749965
#> 2          2.15 0.43821253 0.175573474      0.093475701 0.26543299  4.686368
#> 3          2.15 0.44491739 0.180465027      0.097131387 0.27517446  4.607832
#> 4          2.15 0.45421984 0.185406557      0.100150094 0.28362496  4.551052
#> 5          2.15 0.46282488 0.189705530      0.102655394 0.29039398  4.481283
#> 6          2.15 0.45620201 0.186632449      0.101076541 0.28613490  4.506459
#> 7          2.15 0.45091152 0.183065780      0.098495012 0.27904157  4.559732
#> 8          2.15 0.44949796 0.181434659      0.097070203 0.27528249  4.601891
#> 9          2.15 0.45084347 0.182404642      0.097605302 0.27657828  4.608459
#> 10         2.15 0.46030225 0.187431108      0.100769334 0.28490919  4.548679
#> 11         2.15 0.47725146 0.198304399      0.107673331 0.30291437  4.456959
#> 12         2.15 0.49222922 0.207486276      0.113634952 0.31847846  4.319305
#> 13         2.15 0.50421770 0.215507932      0.119327710 0.33282487  4.050096
#> 14         2.15 0.50841273 0.222261773      0.125106309 0.34708586  3.964155
#> 15         2.15 0.50058166 0.217145389      0.121896335 0.33861137  4.046432
#> 16         2.15 0.48934248 0.211607474      0.118658808 0.32966323  4.112409
#> 17         2.15 0.48774608 0.212263936      0.119765844 0.33217439  4.116267
#> 18         2.15 0.48828117 0.213055162      0.120643164 0.33428381  4.053095
#> 19         2.15 0.48510750 0.212588971      0.121113767 0.33528398  3.975389
#> 20         2.15 0.48203845 0.215249754      0.124400862 0.34324609  3.928387
#> 21         2.15 0.47953415 0.214276793      0.124146527 0.34270808  3.915873
#> 22         2.15 0.48529411 0.212043596      0.121339646 0.33701591  3.930173
#> 23         2.15 0.47607489 0.204438386      0.115935411 0.32320136  3.965661
#> 24         2.15 0.44224748 0.189884646      0.108168519 0.30086587  4.206457
#> 25         2.15 0.42151876 0.180567241      0.102649740 0.28540499  4.415301
#> 26         2.15 0.40240367 0.169469349      0.094929432 0.26489940  4.671371
#> 27         2.15 0.38948954 0.160366974      0.087930219 0.24679614  4.743310
#> 28         2.15 0.37838379 0.152126475      0.081796896 0.23111185  4.828392
#> 29         2.15 0.37593498 0.151931847      0.082105920 0.23178233  4.815238
#> 30         2.15 0.36253041 0.143402674      0.076115331 0.21629141  4.946660
#> 31         2.15 0.35132651 0.136842404      0.071966707 0.20545487  5.023404
#> 32         2.15 0.34322170 0.131694622      0.068522091 0.19671527  5.037383
#> 33         2.15 0.33864194 0.128550324      0.066538588 0.19152274  5.054164
#> 34         2.15 0.33776004 0.128203128      0.066304002 0.19089153  5.062764
#> 35         2.15 0.34086947 0.129860176      0.067280140 0.19351648  5.045060
#> 36         2.15 0.34872619 0.134136291      0.069879552 0.20051865  4.987404
#> 37         2.15 0.35705284 0.137868079      0.072171874 0.20668248  4.903704
#> 38         2.15 0.36190938 0.140180808      0.073663666 0.21059672  4.838839
#> 39         2.15 0.36681789 0.142986516      0.075539118 0.21544714  4.753060
#> 40         2.15 0.38170585 0.151567138      0.081150500 0.23025775  4.529296
#> 41         2.15 0.37867799 0.148938052      0.079414591 0.22571876  4.539248
#> 42         2.15 0.37211527 0.145645223      0.077307330 0.22009038  4.574368
#> 43         2.15 0.36761984 0.142834949      0.075411364 0.21513504  4.559430
#> 44         2.15 0.36875900 0.143565498      0.076030838 0.21675154  4.533114
#> 45         2.15 0.36307844 0.141434976      0.075002113 0.21371013  4.576789
#> 46         2.15 0.12758111 0.049459160      0.027767321 0.07971246 12.685808
#> 47         2.15 0.13347845 0.051361425      0.028699650 0.08247968 12.564294
#> 48         2.15 0.15872419 0.061639716      0.034206873 0.09834164 11.729736
#> 49         2.15 0.15511219 0.059534865      0.033043340 0.09519231 11.816001
#> 50         2.15 0.14200873 0.054793986      0.030590110 0.08810016 11.870391
#> 51         2.15 0.10084309 0.039090730      0.022569992 0.06481986 14.477099
#> 52         2.15 0.12486029 0.049421178      0.028333168 0.08103435 13.143611
#> 53         2.15 0.13173340 0.054270677      0.031846061 0.09027199 12.896922
#> 54         2.15 0.12725453 0.052807886      0.031145174 0.08831886 13.138433
#> 55         2.15 0.13822575 0.055769636      0.031736232 0.09041367 11.991365
#> 56         2.15 0.13265488 0.053644369      0.030939551 0.08795731 11.814562
#> 57         2.15 0.13244217 0.054638398      0.032176633 0.09103167 11.460216
#> 58         2.15 0.13245127 0.054322383      0.031909952 0.09050038 12.228710
#> 59         2.15 0.12736623 0.052118446      0.030683918 0.08734684 12.623060
#> 60         2.15 0.11795163 0.047767824      0.027942746 0.07981077 13.117059
#> 61         2.15 0.13025346 0.055149877      0.033076979 0.09347332 12.952346
#> 62         2.15 0.12907547 0.054294911      0.032318246 0.09130682 13.258961
#> 63         2.15 0.12751906 0.053697954      0.031891261 0.09024368 13.528425
#> 64         2.15 0.12977511 0.054683928      0.032694675 0.09254658 13.076648
#> 65         2.15 0.11849733 0.050652825      0.030758173 0.08677293 13.727282
#> 66         2.15 0.11580028 0.048188424      0.028300881 0.07999753 13.778011
#> 67         2.15 0.10702005 0.041729366      0.023337784 0.06691386 13.615806
#> 68         2.15 0.10377128 0.040878110      0.023128326 0.06606511 13.703200
#> 69         2.15 0.09455180 0.036613282      0.020472345 0.05872706 13.974827
#> 70         2.15 0.09041813 0.035731697      0.020271081 0.05777076 14.768119
#> 71         2.15 0.07459130 0.028837648      0.016080192 0.04603129 15.851153
#> 72         2.15 0.06347560 0.023555592      0.012839418 0.03714898 16.446853
#> 73         2.15 0.05739372 0.021046851      0.011339549 0.03297398 16.995347
#> 74         2.15 0.05495404 0.019698157      0.010455045 0.03062429 16.993047
#> 75         2.15 0.04853226 0.017481788      0.009306949 0.02725444 17.582610
#> 76         2.15 0.04425843 0.016215790      0.008704368 0.02540209 18.312843
#> 77         2.15 0.04116925 0.014668458      0.007687096 0.02268364 19.032441
#> 78         2.15 0.03672052 0.012500528      0.006497476 0.01933329 19.522933
#> 79         2.15 0.03426283 0.011415129      0.005961698 0.01782960 19.527259
#> 80         2.15 0.03166772 0.010249389      0.005080003 0.01541882 19.937569
#> 81         2.15 0.03424041 0.011500410      0.005791601 0.01733837 20.110826
#> 82         2.15 0.03482042 0.011790789      0.005919324 0.01776856 20.262932
#> 83         2.15 0.03432246 0.011716158      0.005899791 0.01763223 20.456772
#> 84         2.15 0.03510453 0.012025347      0.006128680 0.01825649 20.495733
#> 85         2.15 0.03153234 0.011170711      0.006028925 0.01770432 19.391203
#> 86         2.15 0.03679714 0.012645172      0.006375547 0.01902979 19.584810
#> 87         2.15 0.02842836 0.009863431      0.005134792 0.01526176 20.730196
#> 88         2.15 0.02508914 0.008866915      0.004764375 0.01401903 21.393340
#> 89         2.15 0.02471605 0.008811457      0.004744323 0.01392815 21.706600
#> 90         2.15 0.02454394 0.008746876      0.004711686 0.01381392 21.979109
#>    pop_in_poverty       spr        pg estimate_type
#> 1       103169662 0.6065094 15.325681    projection
#> 2       107722286 0.6093514 15.584901    projection
#> 3       112850912 0.6137755 15.950515    projection
#> 4       118759502 0.6218635 16.255413    projection
#> 5       124708352 0.6277353 16.485747    projection
#> 6       126651611 0.6229176 16.367415    projection
#> 7       128991062 0.6182811 16.095038    projection
#> 8       132376532 0.6184546 15.940664    projection
#> 9       136578125 0.6201085 15.964918    projection
#> 10      143498618 0.6272347 16.253345    projection
#> 11      152931872 0.6392894 16.820852    projection
#> 12      161984126 0.6501096 17.330840    projection
#> 13      170588949 0.6568586 17.856235        actual
#> 14      176643840 0.6605470 18.358529        actual
#> 15      178497597 0.6551652 18.078346        actual
#> 16      179167135 0.6489526 17.790233        actual
#> 17      183219977 0.6495245 17.905841    projection
#> 18      188235204 0.6524412 18.020417        actual
#> 19      191981746 0.6522526 18.153056        actual
#> 20      195783128 0.6489951 18.504775        actual
#> 21      199873606 0.6448766 18.500903        actual
#> 22      207618700 0.6524805 18.352193        actual
#> 23      209079373 0.6438843 17.918977        actual
#> 24      199421791 0.6094818 17.140157        actual
#> 25      195195490 0.5931298 16.559836        actual
#> 26      191385683 0.5808191 15.719029        actual
#> 27      190297076 0.5712187 15.008785        actual
#> 28      189975437 0.5631981 14.427666        actual
#> 29      193983748 0.5601640 14.466325        actual
#> 30      192252918 0.5496669 13.915081        actual
#> 31      191380896 0.5405087 13.542848        actual
#> 32      192070281 0.5359166 13.277582        actual
#> 33      194787758 0.5315254 13.104822        actual
#> 34      199605709 0.5295755 13.066818        actual
#> 35      206949786 0.5312359 13.141451        actual
#> 36      217385237 0.5374672 13.377883        actual
#> 37      228534788 0.5422655 13.616662        actual
#> 38      238064380 0.5457029 13.783042        actual
#> 39      247950619 0.5487115 13.983763    projection
#> 40      265074138 0.5610098 14.563503        actual
#> 41      270031839 0.5613019 14.454736        actual
#> 42      272321919 0.5567246 14.244015        actual
#> 43      275900073 0.5540615 14.092019    projection
#> 44      283684442 0.5537737 14.164284       nowcast
#> 45      286417661 0.5483001 14.034081       nowcast
#> 46       46402177 0.3481803  7.497425        actual
#> 47       49602067 0.3517465  7.623677        actual
#> 48       60260331 0.3648431  8.401517        actual
#> 49       60150284 0.3621156  8.305235        actual
#> 50       56224508 0.3572308  7.954611        actual
#> 51       40744886 0.3359948  6.745903        actual
#> 52       51462680 0.3480899  7.529395        actual
#> 53       55369469 0.3478761  7.891745        actual
#> 54       54525277 0.3451173  7.818285        actual
#> 55       60357703 0.3500661  7.929936        actual
#> 56       59003965 0.3456009  7.839854        actual
#> 57       59979387 0.3451453  8.037451        actual
#> 58       61047696 0.3503307  7.986493        actual
#> 59       59717190 0.3487113  7.785102        actual
#> 60       56238940 0.3442966  7.519271        actual
#> 61       63129669 0.3520986  8.084502        actual
#> 62       63567412 0.3519357  7.901111        actual
#> 63       63798876 0.3488033  7.830460        actual
#> 64       65934021 0.3511054  7.994837        actual
#> 65       61096069 0.3451592  7.658184        actual
#> 66       60549405 0.3435557  7.299803        actual
#> 67       56718738 0.3418449  6.792013        actual
#> 68       55716743 0.3367531  6.716480        actual
#> 69       51409369 0.3301083  6.349273        actual
#> 70       49763583 0.3258097  6.155342        actual
#> 71       41547644 0.3147439  5.479650        actual
#> 72       35775180 0.3086224  5.034154        actual
#> 73       32724619 0.3023136  4.777429        actual
#> 74       31690096 0.2985083  4.658420        actual
#> 75       28295965 0.2909722  4.398511        actual
#> 76       26085350 0.2852065  4.218144        actual
#> 77       24526180 0.2845864  4.052175        actual
#> 78       22106488 0.2793164  3.859222        actual
#> 79       20839353 0.2751645  3.795360        actual
#> 80       19452763 0.2722115  3.600245        actual
#> 81       21235652 0.2739847  3.654151        actual
#> 82       21795285 0.2710610  3.624599        actual
#> 83       21671569 0.2692400  3.596558        actual
#> 84       22346802 0.2715972  3.604058        actual
#> 85       20216504 0.2576947  3.641482        actual
#> 86       23730196 0.2687202  3.678749        actual
#> 87       18444458 0.2588591  3.390942        actual
#> 88       16391352 0.2587329  3.269550        actual
#> 89       16260528 0.2582089  3.203427       nowcast
#> 90       16255883 0.2598353  3.174818       nowcast
# }