World Bank country indicator data

List all country indicators supported by the World Bank API.

Usage

wb_country_indicator(
  indicator = "NY.GDP.MKTP.CD",
  country = NULL,
  lang = "en",
  start_date = NULL,
  end_date = NULL
)

Source

https://api.worldbank.org/v2/country/{country}/indicator/{indicator}

Arguments

indicator

(character()) indicators to query.

country

(character()) countries to query. Default NULL. If NULL, all countries are returned.

lang

(character(1)) language to query. Default "en".

start_date

(character(1) | integer(1)) start date to query. Default NULL. Supported formats:

YYYY for yearly data (e.g. 2020 or "2020")

YYYYQ[1-4] for quarterly data (e.g. "2020Q1")

YYYYM[1-12] for monthly data (e.g. "2020M02")

end_date

(character(1) | integer(1)) end date to query, in the same format as start_date. Default NULL.

Value

A data.frame() with the available country indicators. The columns are:

date

The date

indicator_id

The indicator ID.

indicator_name

The indicator name.

country_id

The country ID.

country_name

The country name.

country_code

The country code.

value

The indicator value.

unit

The indicator unit.

obs_status

The observation status.

decimal

The decimal.

Examples

ind <- wb_country_indicator("NY.GDP.MKTP.CD", "US")
head(ind)
#>   date   indicator_id    indicator_name country_id  country_name country_code
#> 1 2023 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 2 2022 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 3 2021 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 4 2020 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 5 2019 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#> 6 2018 NY.GDP.MKTP.CD GDP (current US$)         US United States          USA
#>          value unit obs_status decimal
#> 1 2.772071e+13 <NA>       <NA>       0
#> 2 2.600689e+13 <NA>       <NA>       0
#> 3 2.368117e+13 <NA>       <NA>       0
#> 4 2.135410e+13 <NA>       <NA>       0
#> 5 2.153998e+13 <NA>       <NA>       0
#> 6 2.065652e+13 <NA>       <NA>       0
ind <- wb_country_indicator(
  indicator = c("NY.GDP.MKTP.CD", "FP.CPI.TOTL.ZG"),
  country = c("US", "DE", "FR", "CH", "JP"),
  start_date = "2012M01", end_date = "2012M08"
)
head(ind)
#>   date   indicator_id    indicator_name country_id country_name country_code
#> 1 2023 NY.GDP.MKTP.CD GDP (current US$)         CH  Switzerland          CHE
#> 2 2022 NY.GDP.MKTP.CD GDP (current US$)         CH  Switzerland          CHE
#> 3 2021 NY.GDP.MKTP.CD GDP (current US$)         CH  Switzerland          CHE
#> 4 2020 NY.GDP.MKTP.CD GDP (current US$)         CH  Switzerland          CHE
#> 5 2019 NY.GDP.MKTP.CD GDP (current US$)         CH  Switzerland          CHE
#> 6 2018 NY.GDP.MKTP.CD GDP (current US$)         CH  Switzerland          CHE
#>          value unit obs_status decimal
#> 1 884940402230 <NA>       <NA>       0
#> 2 818426550206 <NA>       <NA>       0
#> 3 813408787222 <NA>       <NA>       0
#> 4 741999406006 <NA>       <NA>       0
#> 5 721369112727 <NA>       <NA>       0
#> 6 725568717468 <NA>       <NA>       0