World Bank topic data

List all topics supported by the World Bank API.

Usage

wb_topic(topic = NULL, lang = "en")

Source

https://api.worldbank.org/v2/topics

Arguments

topic

(NULL | character())
Topic to query. Default NULL. If NULL, all topics are returned.

lang

(character(1))
Language to query. Default "en".

Value

A data.frame() with the available topics. The columns are:

id

The topic ID.

value

The topic value.

source_note

The source note.

See also

Examples

topic <- wb_topic()
head(topic)
#>   id                           value
#> 1  1 Agriculture & Rural Development
#> 2  2               Aid Effectiveness
#> 3  3                Economy & Growth
#> 4  4                       Education
#> 5  5                 Energy & Mining
#> 6  6                     Environment
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            source_note
#> 1                                                                                                                                                        For the 70 percent of the world's poor who live in rural areas, agriculture is the main source of income and employment. But depletion and degradation of land and water pose serious challenges to producing enough food and other agricultural products to sustain livelihoods here and meet the needs of urban populations. Data presented here include measures of agricultural inputs, outputs, and productivity compiled by the UN's Food and Agriculture Organization.
#> 2                                                                                                                                                                                                                                                         Aid effectiveness is the impact that aid has in reducing poverty and inequality, increasing growth, building capacity, and accelerating achievement of the Millennium Development Goals set by the international community. Indicators here cover aid received as well as progress in reducing poverty and improving education, health, and other measures of human welfare.
#> 3        Economic growth is central to economic development. When national income grows, real people benefit. While there is no known formula for stimulating economic growth, data can help policy-makers better understand their countries' economic situations and guide any work toward improvement. Data here covers measures of economic growth, such as gross domestic product (GDP) and gross national income (GNI). It also includes indicators representing factors known to be relevant to economic growth, such as capital stock, employment, investment, savings, consumption, government spending, imports, and exports.
#> 4                                                                                                                                               Education is one of the most powerful instruments for reducing poverty and inequality and lays a foundation for sustained economic growth. The World Bank compiles data on education inputs, participation, efficiency, and outcomes. Data on education are compiled by the United Nations Educational, Scientific, and Cultural Organization (UNESCO) Institute for Statistics from official responses to surveys and from reports provided by education authorities in each country.
#> 5                                                                                                           The world economy needs ever-increasing amounts of energy to sustain economic growth, raise living standards, and reduce poverty. But today's trends in energy use are not sustainable. As the world's population grows and economies become more industrialized, nonrenewable energy sources will become scarcer and more costly. Data here on energy production, use, dependency, and efficiency are compiled by the World Bank from the International Energy Agency and the Carbon Dioxide Information Analysis Center.
#> 6 Natural and man-made environmental resources – fresh water, clean air, forests, grasslands, marine resources, and agro-ecosystems – provide sustenance and a foundation for social and economic development.  The need to safeguard these resources crosses all borders.  Today, the World Bank is one of the key promoters and financiers of environmental upgrading in the developing world. Data here cover forests, biodiversity, emissions, and pollution. Other indicators relevant to the environment are found under data pages for Agriculture & Rural Development, Energy & Mining, Infrastructure, and Urban Development.